chore: split components for docs in docs-src
