+++ docs-src/main.ts
@@ -1,0 +1,16 @@
+import './components/hello-world.ts'
+import './components/my-counter.ts'
+import './components/my-slider.ts'
+import './components/input-button.ts'
+import './components/input-checkbox.ts'
+import './components/input-radiogroup.ts'
+import './components/input-field.ts'
+import './components/code-block.ts'
+import './components/accordion-panel.ts'
+import './components/tab-list.ts'
+import './components/lazy-load.ts'
+import './components/media-context.ts'
+import './components/todo-form.ts'
+import './components/todo-list.ts'
+import './components/todo-count.ts'
+import './components/todo-app.ts'
