+++ docs/assets/js/main.js
@@ -1,1 +1,1 @@
-var Se=(e)=>typeof e==="function",rt=(e)=>Se(e)&&/^async\s+/.test(e.toString()),ye=(e)=>(t)=>t instanceof e,Oe=ye(Error),st=ye(Promise),$e="Computed",ve=(e,t)=>{t=t??rt(e);let a=[],n,r=null,h=!0,m=()=>{if(h=!0,t)Re(a)},f={[Symbol.toStringTag]:$e,get:()=>{if(t)Ce(a);if(!t||h)De(()=>{let b=(S)=>{n=S,h=!1,r=null},A=(S)=>{r=Oe(S)?S:new Error(`Computed function failed: ${S}`)};try{let S=e(n);st(S)?S.then(b).catch(A):b(S)}catch(S){A(S)}},m);if(Oe(r))throw r;return n},map:(b)=>ve(()=>b(f.get()))};return f},it=(e)=>!!e&&typeof e==="object"&&e[Symbol.toStringTag]===$e,H,lt=!1,ot=[],fe=(e)=>me(e)||it(e),Ne=(e,t=!1)=>fe(e)?e:Se(e)?ve(e,t):ut(e),Ce=(e)=>{if(H&&!e.includes(H))e.push(H)},Re=(e)=>e.forEach((t)=>lt?ot.push(t):t()),De=(e,t)=>{let a=H;H=t,e(),H=a};var pe=Symbol();class ke{$;watchers=[];constructor(e){this.value=e}get(){return Ce(this.watchers),this.value}set(e){if(pe!==e){let t=Se(e)?e(this.value):e;if(Object.is(this.value,t))return;this.value=t}if(Re(this.watchers),pe===e)this.watchers=[]}map(e){return ve(()=>e(this.get()))}}var ut=(e)=>new ke(e),me=ye(ke),B=(e)=>{let t=()=>De(()=>{try{e()}catch(a){console.error(a)}},t);t()};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=(e)=>setTimeout(e,16);var re=new Map,Y=[],se,Le=()=>{se=null,Y.forEach((e)=>e()),Y=[],re.clear()},ct=()=>{if(se)cancelAnimationFrame(se);se=requestAnimationFrame(Le)};queueMicrotask(Le);var C=(e,t)=>new Promise((a,n)=>{let r=()=>{try{a(e())}catch(h){n(h)}};if(t){let[h,m]=t;if(!re.has(h))re.set(h,new Map);let f=re.get(h);if(f.has(m)){let b=Y.indexOf(e);if(b>-1)Y.splice(b,1)}f.set(m,r)}Y.push(r),ct()});var dt=(e)=>e.nodeType===Node.COMMENT_NODE,gt=(e)=>!/^on/i.test(e),ht=(e)=>{if(/^(mailto|tel):/i.test(e))return!0;if(e.includes("://"))try{let t=new URL(e,window.location.origin);return!["http:","https:","ftp:"].includes(t.protocol)}catch(t){return!0}return!0},ft=(e,t,a)=>{if(!gt(t))throw new Error(`Unsafe attribute: ${t}`);if(a=String(a).trim(),!ht(a))throw new Error(`Unsafe URL for ${t}: ${a}`);e.setAttribute(t,a)};var pt=(e,t)=>C(()=>{return Array.from(e.childNodes).filter((a)=>!dt(a)).forEach((a)=>a.remove()),e.append(document.createTextNode(t)),e},[e,"t"]),mt=(e,t,a)=>C(()=>{return ft(e,t,a),e},[e,`a:${t}`]),bt=(e,t)=>C(()=>{return e.removeAttribute(t),e},[e,`a:${t}`]),St=(e,t,a)=>C(()=>{return e.toggleAttribute(t,a),e},[e,`a:${t}`]),yt=(e,t,a)=>C(()=>{return e.classList.toggle(t,a),e},[e,`c:${t}`]);var L=(e)=>typeof e==="function",Ge=(e)=>!!e&&typeof e==="object",vt=(e)=>typeof e==="number",le=(e)=>typeof e==="string",kt=(e)=>typeof e==="symbol",Ft=(e)=>le(e)||kt(e)||vt(e),x=!1,wt="debug";var At="warn",X="error",xt=(e)=>e?`#${e}`:"",Et=(e)=>e.length?`.${Array.from(e).join(".")}`:"",J=(e)=>`<${e.localName}${xt(e.id)}${Et(e.classList)}>`,D=(e)=>le(e)?`"${e}"`:Ge(e)?JSON.stringify(e):String(e),E=(e,t,a=wt)=>{if(x||[X,At].includes(a))console[a](t,e);return e},_t=(e)=>L(e)&&e.length===2,he=(e,t,a=0)=>_t(e)?e(t,a):e;class ie{$;A;constructor(e,t=[e]){this.host=e,this.targets=t}on(e,t){return this.targets.forEach((a,n)=>a.addEventListener(e,he(t,a,n))),this}off(e,t){return this.targets.forEach((a,n)=>a.removeEventListener(e,he(t,a,n))),this}pass(e){return this.targets.forEach(async(t,a)=>{if(await y.registry.whenDefined(t.localName),t instanceof y)Object.entries(e).forEach(([n,r])=>{let h=he(r,t,a),m=Ft(h)?this.host.signals.get(h):Ne(h,!0);if(m)t.set(n,m);else E(r,`Invalid source for state ${D(n)}`,X)});else E(t,"Target is not a UIElement",X)}),this}sync(...e){return this.targets.forEach((t,a)=>e.forEach((n)=>n(this.host,t,a))),this}}var Ot=(e,t)=>{if(t==null)return;let a=e(t);return Number.isFinite(a)?a:void 0},be=(e,t,a,n=void 0)=>{let r=e.constructor.states[t];return L(r)&&!!r.length?r(a,e,n):a},I=(e)=>e!=null,je=(e)=>Ot(parseInt,e);var Te="context-request";class Pe extends Event{$;A;B;constructor(e,t,a=!1){super(Te,{bubbles:!0,composed:!0});this.context=e,this.callback=t,this.subscribe=a}}var $t=(e)=>{let t=e.constructor,a=t.consumedContexts||[];setTimeout(()=>{for(let r of a)e.dispatchEvent(new Pe(r,(h)=>e.set(String(r),h)))});let n=t.providedContexts||[];if(!n.length)return!1;return e.addEventListener(Te,(r)=>{let{context:h,callback:m}=r;if(!n.includes(h)||!L(m))return;r.stopPropagation(),m(e.signals.get(String(h)))}),!0};class y extends HTMLElement{static registry=customElements;static states={};static observedAttributes;static consumedContexts;static providedContexts;static define(e){try{if(y.registry.define(e,this),x)E(e,"Registered custom element")}catch(t){E(t,`Failed to register custom element ${e}`,X)}}signals=new Map;self=new ie(this);root=this.shadowRoot||this;debug=!1;attributeChangedCallback(e,t,a){if(a===t)return;if(x&&this.debug)E(`${D(t)} => ${D(a)}`,`Attribute "${e}" of ${J(this)} changed`);this.set(e,be(this,e,a,t))}connectedCallback(){if(x){if(this.debug=this.hasAttribute("debug"),this.debug)E(this,"Connected")}Object.entries(this.constructor.states).forEach(([e,t])=>{let a=L(t)?be(this,e,this.getAttribute(e)??void 0,void 0):t;this.set(e,a,!1)}),$t(this)}disconnectedCallback(){if(x&&this.debug)E(this,"Disconnected")}adoptedCallback(){if(x&&this.debug)E(this,"Adopted")}has(e){return this.signals.has(e)}get(e){let t=(n)=>!Ge(n)?n:L(n)?t(n()):fe(n)?t(n.get()):n,a=t(this.signals.get(e));if(x&&this.debug)E(a,`Get current value of state ${D(e)} in ${J(this)}`);return a}set(e,t,a=!0){let n;if(!this.signals.has(e)){if(x&&this.debug)n="Create";this.signals.set(e,Ne(t,!0))}else if(a){let r=this.signals.get(e);if(fe(t)){if(x&&this.debug)n="Replace";if(this.signals.set(e,t),me(r))r.set(pe)}else if(me(r)){if(x&&this.debug)n="Update";r.set(t)}else{E(t,`Computed state ${D(e)} in ${J(this)} cannot be set`,X);return}}else return;if(x&&this.debug)E(t,`${n} state ${D(e)} in ${J(this)}`)}delete(e){if(x&&this.debug)E(e,`Delete state ${D(e)} from ${J(this)}`);return this.signals.delete(e)}first(e){let t=this.root.querySelector(e);return new ie(this,t?[t]:[])}all(e){return new ie(this,Array.from(this.root.querySelectorAll(e)))}}var Q=(e,t)=>(a,n)=>{let{read:r,update:h}=t,m=r(n);if(!L(e)){let f=le(e)&&le(m)?be(a,e,m):m;a.set(e,f,!1)}B(()=>{let f=r(n),b=L(e)?e(f):a.get(e);if(!Object.is(b,f)){if(b===null&&t.delete)t.delete(n);else if(b==null&&m)h(n,m);else if(b!=null)h(n,b)}})};var $=(e)=>Q(e,{read:(t)=>t.textContent,update:(t,a)=>pt(t,a)}),F=(e,t=e)=>Q(t,{read:(a)=>a[e],update:(a,n)=>a[e]=n}),W=(e,t=e)=>Q(t,{read:(a)=>a.getAttribute(e),update:(a,n)=>mt(a,e,n),delete:(a)=>bt(a,e)}),U=(e,t=e)=>Q(t,{read:(a)=>a.hasAttribute(e),update:(a,n)=>St(a,e,n)}),M=(e,t=e)=>Q(t,{read:(a)=>a.classList.contains(e),update:(a,n)=>yt(a,e,n)});var Nt=typeof window!="undefined"?window:typeof WorkerGlobalScope!="undefined"&&self instanceof WorkerGlobalScope?self:{},d=function(e){var t=/(?:^|\s)lang(?:uage)?-([\w-]+)(?=\s|$)/i,a=0,n={},r={manual:e.Prism&&e.Prism.manual,disableWorkerMessageHandler:e.Prism&&e.Prism.disableWorkerMessageHandler,util:{encode:function i(s){return s instanceof h?new h(s.type,i(s.content),s.alias):Array.isArray(s)?s.map(i):s.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/\u00a0/g," ")},type:function(i){return Object.prototype.toString.call(i).slice(8,-1)},objId:function(i){return i.__id||Object.defineProperty(i,"__id",{value:++a}),i.__id},clone:function i(s,u){var l,o;switch(u=u||{},r.util.type(s)){case"Object":if(o=r.util.objId(s),u[o])return u[o];for(var c in l={},u[o]=l,s)s.hasOwnProperty(c)&&(l[c]=i(s[c],u));return l;case"Array":return o=r.util.objId(s),u[o]?u[o]:(l=[],u[o]=l,s.forEach(function(g,p){l[p]=i(g,u)}),l);default:return s}},getLanguage:function(i){for(;i;){var s=t.exec(i.className);if(s)return s[1].toLowerCase();i=i.parentElement}return"none"},setLanguage:function(i,s){i.className=i.className.replace(RegExp(t,"gi"),""),i.classList.add("language-"+s)},currentScript:function(){if(typeof document=="undefined")return null;if("currentScript"in document)return document.currentScript;try{throw new Error}catch(l){var i=(/at [^(\r\n]*\((.*):[^:]+:[^:]+\)$/i.exec(l.stack)||[])[1];if(i){var s=document.getElementsByTagName("script");for(var u in s)if(s[u].src==i)return s[u]}return null}},isActive:function(i,s,u){for(var l="no-"+s;i;){var o=i.classList;if(o.contains(s))return!0;if(o.contains(l))return!1;i=i.parentElement}return!!u}},languages:{plain:n,plaintext:n,text:n,txt:n,extend:function(i,s){var u=r.util.clone(r.languages[i]);for(var l in s)u[l]=s[l];return u},insertBefore:function(i,s,u,l){var o=(l=l||r.languages)[i],c={};for(var g in o)if(o.hasOwnProperty(g)){if(g==s)for(var p in u)u.hasOwnProperty(p)&&(c[p]=u[p]);u.hasOwnProperty(g)||(c[g]=o[g])}var k=l[i];return l[i]=c,r.languages.DFS(r.languages,function(v,j){j===k&&v!=i&&(this[v]=c)}),c},DFS:function i(s,u,l,o){o=o||{};var c=r.util.objId;for(var g in s)if(s.hasOwnProperty(g)){u.call(s,g,s[g],l||g);var p=s[g],k=r.util.type(p);k!=="Object"||o[c(p)]?k!=="Array"||o[c(p)]||(o[c(p)]=!0,i(p,u,g,o)):(o[c(p)]=!0,i(p,u,null,o))}}},plugins:{},highlightAll:function(i,s){r.highlightAllUnder(document,i,s)},highlightAllUnder:function(i,s,u){var l={callback:u,container:i,selector:'code[class*="language-"], [class*="language-"] code, code[class*="lang-"], [class*="lang-"] code'};r.hooks.run("before-highlightall",l),l.elements=Array.prototype.slice.apply(l.container.querySelectorAll(l.selector)),r.hooks.run("before-all-elements-highlight",l);for(var o,c=0;o=l.elements[c++];)r.highlightElement(o,s===!0,l.callback)},highlightElement:function(i,s,u){var l=r.util.getLanguage(i),o=r.languages[l];r.util.setLanguage(i,l);var c=i.parentElement;c&&c.nodeName.toLowerCase()==="pre"&&r.util.setLanguage(c,l);var g={element:i,language:l,grammar:o,code:i.textContent};function p(v){g.highlightedCode=v,r.hooks.run("before-insert",g),g.element.innerHTML=g.highlightedCode,r.hooks.run("after-highlight",g),r.hooks.run("complete",g),u&&u.call(g.element)}if(r.hooks.run("before-sanity-check",g),(c=g.element.parentElement)&&c.nodeName.toLowerCase()==="pre"&&!c.hasAttribute("tabindex")&&c.setAttribute("tabindex","0"),!g.code)return r.hooks.run("complete",g),void(u&&u.call(g.element));if(r.hooks.run("before-highlight",g),g.grammar)if(s&&e.Worker){var k=new Worker(r.filename);k.onmessage=function(v){p(v.data)},k.postMessage(JSON.stringify({language:g.language,code:g.code,immediateClose:!0}))}else p(r.highlight(g.code,g.grammar,g.language));else p(r.util.encode(g.code))},highlight:function(i,s,u){var l={code:i,grammar:s,language:u};if(r.hooks.run("before-tokenize",l),!l.grammar)throw new Error('The language "'+l.language+'" has no grammar.');return l.tokens=r.tokenize(l.code,l.grammar),r.hooks.run("after-tokenize",l),h.stringify(r.util.encode(l.tokens),l.language)},tokenize:function(i,s){var u=s.rest;if(u){for(var l in u)s[l]=u[l];delete s.rest}var o=new b;return A(o,o.head,i),f(i,o,s,o.head,0),function(c){for(var g=[],p=c.head.next;p!==c.tail;)g.push(p.value),p=p.next;return g}(o)},hooks:{all:{},add:function(i,s){var u=r.hooks.all;u[i]=u[i]||[],u[i].push(s)},run:function(i,s){var u=r.hooks.all[i];if(u&&u.length)for(var l,o=0;l=u[o++];)l(s)}},Token:h};function h(i,s,u,l){this.type=i,this.content=s,this.alias=u,this.length=0|(l||"").length}function m(i,s,u,l){i.lastIndex=s;var o=i.exec(u);if(o&&l&&o[1]){var c=o[1].length;o.index+=c,o[0]=o[0].slice(c)}return o}function f(i,s,u,l,o,c){for(var g in u)if(u.hasOwnProperty(g)&&u[g]){var p=u[g];p=Array.isArray(p)?p:[p];for(var k=0;k<p.length;++k){if(c&&c.cause==g+","+k)return;var v=p[k],j=v.inside,T=!!v.lookbehind,K=!!v.greedy,V=v.alias;if(K&&!v.pattern.global){var at=v.pattern.toString().match(/[imsuy]*$/)[0];v.pattern=RegExp(v.pattern.source,at+"g")}for(var Ee=v.pattern||v,w=l.next,_=o;w!==s.tail&&!(c&&_>=c.reach);_+=w.value.length,w=w.next){var P=w.value;if(s.length>i.length)return;if(!(P instanceof h)){var O,ee=1;if(K){if(!(O=m(Ee,_,i,T))||O.index>=i.length)break;var te=O.index,nt=O.index+O[0].length,R=_;for(R+=w.value.length;te>=R;)R+=(w=w.next).value.length;if(_=R-=w.value.length,w.value instanceof h)continue;for(var Z=w;Z!==s.tail&&(R<nt||typeof Z.value=="string");Z=Z.next)ee++,R+=Z.value.length;ee--,P=i.slice(_,R),O.index-=_}else if(!(O=m(Ee,0,P,T)))continue;te=O.index;var ae=O[0],ce=P.slice(0,te),_e=P.slice(te+ae.length),de=_+P.length;c&&de>c.reach&&(c.reach=de);var ne=w.prev;if(ce&&(ne=A(s,ne,ce),_+=ce.length),S(s,ne,ee),w=A(s,ne,new h(g,j?r.tokenize(ae,j):ae,V,ae)),_e&&A(s,w,_e),ee>1){var ge={cause:g+","+k,reach:de};f(i,s,u,w.prev,_,ge),c&&ge.reach>c.reach&&(c.reach=ge.reach)}}}}}}function b(){var i={value:null,prev:null,next:null},s={value:null,prev:i,next:null};i.next=s,this.head=i,this.tail=s,this.length=0}function A(i,s,u){var l=s.next,o={value:u,prev:s,next:l};return s.next=o,l.prev=o,i.length++,o}function S(i,s,u){for(var l=s.next,o=0;o<u&&l!==i.tail;o++)l=l.next;s.next=l,l.prev=s,i.length-=o}if(e.Prism=r,h.stringify=function i(s,u){if(typeof s=="string")return s;if(Array.isArray(s)){var l="";return s.forEach(function(k){l+=i(k,u)}),l}var o={type:s.type,content:i(s.content,u),tag:"span",classes:["token",s.type],attributes:{},language:u},c=s.alias;c&&(Array.isArray(c)?Array.prototype.push.apply(o.classes,c):o.classes.push(c)),r.hooks.run("wrap",o);var g="";for(var p in o.attributes)g+=" "+p+'="'+(o.attributes[p]||"").replace(/"/g,"&quot;")+'"';return"<"+o.tag+' class="'+o.classes.join(" ")+'"'+g+">"+o.content+"</"+o.tag+">"},!e.document)return e.addEventListener?(r.disableWorkerMessageHandler||e.addEventListener("message",function(i){var s=JSON.parse(i.data),u=s.language,l=s.code,o=s.immediateClose;e.postMessage(r.highlight(l,r.languages[u],u)),o&&e.close()},!1),r):r;var N=r.util.currentScript();function G(){r.manual||r.highlightAll()}if(N&&(r.filename=N.src,N.hasAttribute("data-manual")&&(r.manual=!0)),!r.manual){var q=document.readyState;q==="loading"||q==="interactive"&&N&&N.defer?document.addEventListener("DOMContentLoaded",G):window.requestAnimationFrame?window.requestAnimationFrame(G):window.setTimeout(G,16)}return r}(Nt);typeof Fe!="undefined"&&Fe.exports&&(Fe.exports=d),typeof global!="undefined"&&(global.Prism=d);d.languages.markup={comment:{pattern:/<!--(?:(?!<!--)[\s\S])*?-->/,greedy:!0},prolog:{pattern:/<\?[\s\S]+?\?>/,greedy:!0},doctype:{pattern:/<!DOCTYPE(?:[^>"'[\]]|"[^"]*"|'[^']*')+(?:\[(?:[^<"'\]]|"[^"]*"|'[^']*'|<(?!!--)|<!--(?:[^-]|-(?!->))*-->)*\]\s*)?>/i,greedy:!0,inside:{"internal-subset":{pattern:/(^[^\[]*\[)[\s\S]+(?=\]>$)/,lookbehind:!0,greedy:!0,inside:null},string:{pattern:/"[^"]*"|'[^']*'/,greedy:!0},punctuation:/^<!|>$|[[\]]/,"doctype-tag":/^DOCTYPE/i,name:/[^\s<>'"]+/}},cdata:{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,greedy:!0},tag:{pattern:/<\/?(?!\d)[^\s>\/=$<%]+(?:\s(?:\s*[^\s>\/=]+(?:\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+(?=[\s>]))|(?=[\s/>])))+)?\s*\/?>/,greedy:!0,inside:{tag:{pattern:/^<\/?[^\s>\/]+/,inside:{punctuation:/^<\/?/,namespace:/^[^\s>\/:]+:/}},"special-attr":[],"attr-value":{pattern:/=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+)/,inside:{punctuation:[{pattern:/^=/,alias:"attr-equals"},{pattern:/^(\s*)["']|["']$/,lookbehind:!0}]}},punctuation:/\/?>/,"attr-name":{pattern:/[^\s>\/]+/,inside:{namespace:/^[^\s>\/:]+:/}}}},entity:[{pattern:/&[\da-z]{1,8};/i,alias:"named-entity"},/&#x?[\da-f]{1,8};/i]},d.languages.markup.tag.inside["attr-value"].inside.entity=d.languages.markup.entity,d.languages.markup.doctype.inside["internal-subset"].inside=d.languages.markup,d.hooks.add("wrap",function(e){e.type==="entity"&&(e.attributes.title=e.content.replace(/&amp;/,"&"))}),Object.defineProperty(d.languages.markup.tag,"addInlined",{value:function(e,t){var a={};a["language-"+t]={pattern:/(^<!\[CDATA\[)[\s\S]+?(?=\]\]>$)/i,lookbehind:!0,inside:d.languages[t]},a.cdata=/^<!\[CDATA\[|\]\]>$/i;var n={"included-cdata":{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,inside:a}};n["language-"+t]={pattern:/[\s\S]+/,inside:d.languages[t]};var r={};r[e]={pattern:RegExp("(<__[^>]*>)(?:<!\\[CDATA\\[(?:[^\\]]|\\](?!\\]>))*\\]\\]>|(?!<!\\[CDATA\\[)[^])*?(?=</__>)".replace(/__/g,function(){return e}),"i"),lookbehind:!0,greedy:!0,inside:n},d.languages.insertBefore("markup","cdata",r)}}),Object.defineProperty(d.languages.markup.tag,"addAttribute",{value:function(e,t){d.languages.markup.tag.inside["special-attr"].push({pattern:RegExp("(^|[\"'\\s])(?:"+e+")\\s*=\\s*(?:\"[^\"]*\"|'[^']*'|[^\\s'\">=]+(?=[\\s>]))","i"),lookbehind:!0,inside:{"attr-name":/^[^\s=]+/,"attr-value":{pattern:/=[\s\S]+/,inside:{value:{pattern:/(^=\s*(["']|(?!["'])))\S[\s\S]*(?=\2$)/,lookbehind:!0,alias:[t,"language-"+t],inside:d.languages[t]},punctuation:[{pattern:/^=/,alias:"attr-equals"},/"|'/]}}}})}}),d.languages.html=d.languages.markup,d.languages.mathml=d.languages.markup,d.languages.svg=d.languages.markup,d.languages.xml=d.languages.extend("markup",{}),d.languages.ssml=d.languages.xml,d.languages.atom=d.languages.xml,d.languages.rss=d.languages.xml;(function(e){var t=/(?:"(?:\\(?:\r\n|[\s\S])|[^"\\\r\n])*"|'(?:\\(?:\r\n|[\s\S])|[^'\\\r\n])*')/;e.languages.css={comment:/\/\*[\s\S]*?\*\//,atrule:{pattern:RegExp("@[\\w-](?:[^;{\\s\"']|\\s+(?!\\s)|"+t.source+")*?(?:;|(?=\\s*\\{))"),inside:{rule:/^@[\w-]+/,"selector-function-argument":{pattern:/(\bselector\s*\(\s*(?![\s)]))(?:[^()\s]|\s+(?![\s)])|\((?:[^()]|\([^()]*\))*\))+(?=\s*\))/,lookbehind:!0,alias:"selector"},keyword:{pattern:/(^|[^\w-])(?:and|not|only|or)(?![\w-])/,lookbehind:!0}}},url:{pattern:RegExp("\\burl\\((?:"+t.source+"|(?:[^\\\\\r\n()\"']|\\\\[^])*)\\)","i"),greedy:!0,inside:{function:/^url/i,punctuation:/^\(|\)$/,string:{pattern:RegExp("^"+t.source+"$"),alias:"url"}}},selector:{pattern:RegExp("(^|[{}\\s])[^{}\\s](?:[^{};\"'\\s]|\\s+(?![\\s{])|"+t.source+")*(?=\\s*\\{)"),lookbehind:!0},string:{pattern:t,greedy:!0},property:{pattern:/(^|[^-\w\xA0-\uFFFF])(?!\s)[-_a-z\xA0-\uFFFF](?:(?!\s)[-\w\xA0-\uFFFF])*(?=\s*:)/i,lookbehind:!0},important:/!important\b/i,function:{pattern:/(^|[^-a-z0-9])[-a-z0-9]+(?=\()/i,lookbehind:!0},punctuation:/[(){};:,]/},e.languages.css.atrule.inside.rest=e.languages.css;var a=e.languages.markup;a&&(a.tag.addInlined("style","css"),a.tag.addAttribute("style","css"))})(d);d.languages.clike={comment:[{pattern:/(^|[^\\])\/\*[\s\S]*?(?:\*\/|$)/,lookbehind:!0,greedy:!0},{pattern:/(^|[^\\:])\/\/.*/,lookbehind:!0,greedy:!0}],string:{pattern:/(["'])(?:\\(?:\r\n|[\s\S])|(?!\1)[^\\\r\n])*\1/,greedy:!0},"class-name":{pattern:/(\b(?:class|extends|implements|instanceof|interface|new|trait)\s+|\bcatch\s+\()[\w.\\]+/i,lookbehind:!0,inside:{punctuation:/[.\\]/}},keyword:/\b(?:break|catch|continue|do|else|finally|for|function|if|in|instanceof|new|null|return|throw|try|while)\b/,boolean:/\b(?:false|true)\b/,function:/\b\w+(?=\()/,number:/\b0x[\da-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:e[+-]?\d+)?/i,operator:/[<>]=?|[!=]=?=?|--?|\+\+?|&&?|\|\|?|[?*/~^%]/,punctuation:/[{}[\];(),.:]/};d.languages.javascript=d.languages.extend("clike",{"class-name":[d.languages.clike["class-name"],{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$A-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\.(?:constructor|prototype))/,lookbehind:!0}],keyword:[{pattern:/((?:^|\})\s*)catch\b/,lookbehind:!0},{pattern:/(^|[^.]|\.\.\.\s*)\b(?:as|assert(?=\s*\{)|async(?=\s*(?:function\b|\(|[$\w\xA0-\uFFFF]|$))|await|break|case|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally(?=\s*(?:\{|$))|for|from(?=\s*(?:['"]|$))|function|(?:get|set)(?=\s*(?:[#\[$\w\xA0-\uFFFF]|$))|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)\b/,lookbehind:!0}],function:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*(?:\.\s*(?:apply|bind|call)\s*)?\()/,number:{pattern:RegExp("(^|[^\\w$])(?:NaN|Infinity|0[bB][01]+(?:_[01]+)*n?|0[oO][0-7]+(?:_[0-7]+)*n?|0[xX][\\dA-Fa-f]+(?:_[\\dA-Fa-f]+)*n?|\\d+(?:_\\d+)*n|(?:\\d+(?:_\\d+)*(?:\\.(?:\\d+(?:_\\d+)*)?)?|\\.\\d+(?:_\\d+)*)(?:[Ee][+-]?\\d+(?:_\\d+)*)?)(?![\\w$])"),lookbehind:!0},operator:/--|\+\+|\*\*=?|=>|&&=?|\|\|=?|[!=]==|<<=?|>>>?=?|[-+*/%&|^!=<>]=?|\.{3}|\?\?=?|\?\.?|[~:]/}),d.languages.javascript["class-name"][0].pattern=/(\b(?:class|extends|implements|instanceof|interface|new)\s+)[\w.\\]+/,d.languages.insertBefore("javascript","keyword",{regex:{pattern:RegExp("((?:^|[^$\\w\\xA0-\\uFFFF.\"'\\])\\s]|\\b(?:return|yield))\\s*)/(?:(?:\\[(?:[^\\]\\\\\r\n]|\\\\.)*\\]|\\\\.|[^/\\\\\\[\r\n])+/[dgimyus]{0,7}|(?:\\[(?:[^[\\]\\\\\r\n]|\\\\.|\\[(?:[^[\\]\\\\\r\n]|\\\\.|\\[(?:[^[\\]\\\\\r\n]|\\\\.)*\\])*\\])*\\]|\\\\.|[^/\\\\\\[\r\n])+/[dgimyus]{0,7}v[dgimyus]{0,7})(?=(?:\\s|/\\*(?:[^*]|\\*(?!/))*\\*/)*(?:$|[\r\n,.;:})\\]]|//))"),lookbehind:!0,greedy:!0,inside:{"regex-source":{pattern:/^(\/)[\s\S]+(?=\/[a-z]*$)/,lookbehind:!0,alias:"language-regex",inside:d.languages.regex},"regex-delimiter":/^\/|\/$/,"regex-flags":/^[a-z]+$/}},"function-variable":{pattern:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*[=:]\s*(?:async\s*)?(?:\bfunction\b|(?:\((?:[^()]|\([^()]*\))*\)|(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)\s*=>))/,alias:"function"},parameter:[{pattern:/(function(?:\s+(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)?\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\))/,lookbehind:!0,inside:d.languages.javascript},{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$a-z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*=>)/i,lookbehind:!0,inside:d.languages.javascript},{pattern:/(\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*=>)/,lookbehind:!0,inside:d.languages.javascript},{pattern:/((?:\b|\s|^)(?!(?:as|async|await|break|case|catch|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally|for|from|function|get|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|set|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)(?![$\w\xA0-\uFFFF]))(?:(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*\s*)\(\s*|\]\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*\{)/,lookbehind:!0,inside:d.languages.javascript}],constant:/\b[A-Z](?:[A-Z_]|\dx?)*\b/}),d.languages.insertBefore("javascript","string",{hashbang:{pattern:/^#!.*/,greedy:!0,alias:"comment"},"template-string":{pattern:/`(?:\\[\s\S]|\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}|(?!\$\{)[^\\`])*`/,greedy:!0,inside:{"template-punctuation":{pattern:/^`|`$/,alias:"string"},interpolation:{pattern:/((?:^|[^\\])(?:\\{2})*)\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}/,lookbehind:!0,inside:{"interpolation-punctuation":{pattern:/^\$\{|\}$/,alias:"punctuation"},rest:d.languages.javascript}},string:/[\s\S]+/}},"string-property":{pattern:/((?:^|[,{])[ \t]*)(["'])(?:\\(?:\r\n|[\s\S])|(?!\2)[^\\\r\n])*\2(?=\s*:)/m,lookbehind:!0,greedy:!0,alias:"property"}}),d.languages.insertBefore("javascript","operator",{"literal-property":{pattern:/((?:^|[,{])[ \t]*)(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*:)/m,lookbehind:!0,alias:"property"}}),d.languages.markup&&(d.languages.markup.tag.addInlined("script","javascript"),d.languages.markup.tag.addAttribute("on(?:abort|blur|change|click|composition(?:end|start|update)|dblclick|error|focus(?:in|out)?|key(?:down|up)|load|mouse(?:down|enter|leave|move|out|over|up)|reset|resize|scroll|select|slotchange|submit|unload|wheel)","javascript")),d.languages.js=d.languages.javascript;(function(e){var t="\\b(?:BASH|BASHOPTS|BASH_ALIASES|BASH_ARGC|BASH_ARGV|BASH_CMDS|BASH_COMPLETION_COMPAT_DIR|BASH_LINENO|BASH_REMATCH|BASH_SOURCE|BASH_VERSINFO|BASH_VERSION|COLORTERM|COLUMNS|COMP_WORDBREAKS|DBUS_SESSION_BUS_ADDRESS|DEFAULTS_PATH|DESKTOP_SESSION|DIRSTACK|DISPLAY|EUID|GDMSESSION|GDM_LANG|GNOME_KEYRING_CONTROL|GNOME_KEYRING_PID|GPG_AGENT_INFO|GROUPS|HISTCONTROL|HISTFILE|HISTFILESIZE|HISTSIZE|HOME|HOSTNAME|HOSTTYPE|IFS|INSTANCE|JOB|LANG|LANGUAGE|LC_ADDRESS|LC_ALL|LC_IDENTIFICATION|LC_MEASUREMENT|LC_MONETARY|LC_NAME|LC_NUMERIC|LC_PAPER|LC_TELEPHONE|LC_TIME|LESSCLOSE|LESSOPEN|LINES|LOGNAME|LS_COLORS|MACHTYPE|MAILCHECK|MANDATORY_PATH|NO_AT_BRIDGE|OLDPWD|OPTERR|OPTIND|ORBIT_SOCKETDIR|OSTYPE|PAPERSIZE|PATH|PIPESTATUS|PPID|PS1|PS2|PS3|PS4|PWD|RANDOM|REPLY|SECONDS|SELINUX_INIT|SESSION|SESSIONTYPE|SESSION_MANAGER|SHELL|SHELLOPTS|SHLVL|SSH_AUTH_SOCK|TERM|UID|UPSTART_EVENTS|UPSTART_INSTANCE|UPSTART_JOB|UPSTART_SESSION|USER|WINDOWID|XAUTHORITY|XDG_CONFIG_DIRS|XDG_CURRENT_DESKTOP|XDG_DATA_DIRS|XDG_GREETER_DATA_DIR|XDG_MENU_PREFIX|XDG_RUNTIME_DIR|XDG_SEAT|XDG_SEAT_PATH|XDG_SESSION_DESKTOP|XDG_SESSION_ID|XDG_SESSION_PATH|XDG_SESSION_TYPE|XDG_VTNR|XMODIFIERS)\\b",a={pattern:/(^(["']?)\w+\2)[ \t]+\S.*/,lookbehind:!0,alias:"punctuation",inside:null},n={bash:a,environment:{pattern:RegExp("\\$"+t),alias:"constant"},variable:[{pattern:/\$?\(\([\s\S]+?\)\)/,greedy:!0,inside:{variable:[{pattern:/(^\$\(\([\s\S]+)\)\)/,lookbehind:!0},/^\$\(\(/],number:/\b0x[\dA-Fa-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:[Ee]-?\d+)?/,operator:/--|\+\+|\*\*=?|<<=?|>>=?|&&|\|\||[=!+\-*/%<>^&|]=?|[?~:]/,punctuation:/\(\(?|\)\)?|,|;/}},{pattern:/\$\((?:\([^)]+\)|[^()])+\)|`[^`]+`/,greedy:!0,inside:{variable:/^\$\(|^`|\)$|`$/}},{pattern:/\$\{[^}]+\}/,greedy:!0,inside:{operator:/:[-=?+]?|[!\/]|##?|%%?|\^\^?|,,?/,punctuation:/[\[\]]/,environment:{pattern:RegExp("(\\{)"+t),lookbehind:!0,alias:"constant"}}},/\$(?:\w+|[#?*!@$])/],entity:/\\(?:[abceEfnrtv\\"]|O?[0-7]{1,3}|U[0-9a-fA-F]{8}|u[0-9a-fA-F]{4}|x[0-9a-fA-F]{1,2})/};e.languages.bash={shebang:{pattern:/^#!\s*\/.*/,alias:"important"},comment:{pattern:/(^|[^"{\\$])#.*/,lookbehind:!0},"function-name":[{pattern:/(\bfunction\s+)[\w-]+(?=(?:\s*\(?:\s*\))?\s*\{)/,lookbehind:!0,alias:"function"},{pattern:/\b[\w-]+(?=\s*\(\s*\)\s*\{)/,alias:"function"}],"for-or-select":{pattern:/(\b(?:for|select)\s+)\w+(?=\s+in\s)/,alias:"variable",lookbehind:!0},"assign-left":{pattern:/(^|[\s;|&]|[<>]\()\w+(?:\.\w+)*(?=\+?=)/,inside:{environment:{pattern:RegExp("(^|[\\s;|&]|[<>]\\()"+t),lookbehind:!0,alias:"constant"}},alias:"variable",lookbehind:!0},parameter:{pattern:/(^|\s)-{1,2}(?:\w+:[+-]?)?\w+(?:\.\w+)*(?=[=\s]|$)/,alias:"variable",lookbehind:!0},string:[{pattern:/((?:^|[^<])<<-?\s*)(\w+)\s[\s\S]*?(?:\r?\n|\r)\2/,lookbehind:!0,greedy:!0,inside:n},{pattern:/((?:^|[^<])<<-?\s*)(["'])(\w+)\2\s[\s\S]*?(?:\r?\n|\r)\3/,lookbehind:!0,greedy:!0,inside:{bash:a}},{pattern:/(^|[^\\](?:\\\\)*)"(?:\\[\s\S]|\$\([^)]+\)|\$(?!\()|`[^`]+`|[^"\\`$])*"/,lookbehind:!0,greedy:!0,inside:n},{pattern:/(^|[^$\\])'[^']*'/,lookbehind:!0,greedy:!0},{pattern:/\$'(?:[^'\\]|\\[\s\S])*'/,greedy:!0,inside:{entity:n.entity}}],environment:{pattern:RegExp("\\$?"+t),alias:"constant"},variable:n.variable,function:{pattern:/(^|[\s;|&]|[<>]\()(?:add|apropos|apt|apt-cache|apt-get|aptitude|aspell|automysqlbackup|awk|basename|bash|bc|bconsole|bg|bzip2|cal|cargo|cat|cfdisk|chgrp|chkconfig|chmod|chown|chroot|cksum|clear|cmp|column|comm|composer|cp|cron|crontab|csplit|curl|cut|date|dc|dd|ddrescue|debootstrap|df|diff|diff3|dig|dir|dircolors|dirname|dirs|dmesg|docker|docker-compose|du|egrep|eject|env|ethtool|expand|expect|expr|fdformat|fdisk|fg|fgrep|file|find|fmt|fold|format|free|fsck|ftp|fuser|gawk|git|gparted|grep|groupadd|groupdel|groupmod|groups|grub-mkconfig|gzip|halt|head|hg|history|host|hostname|htop|iconv|id|ifconfig|ifdown|ifup|import|install|ip|java|jobs|join|kill|killall|less|link|ln|locate|logname|logrotate|look|lpc|lpr|lprint|lprintd|lprintq|lprm|ls|lsof|lynx|make|man|mc|mdadm|mkconfig|mkdir|mke2fs|mkfifo|mkfs|mkisofs|mknod|mkswap|mmv|more|most|mount|mtools|mtr|mutt|mv|nano|nc|netstat|nice|nl|node|nohup|notify-send|npm|nslookup|op|open|parted|passwd|paste|pathchk|ping|pkill|pnpm|podman|podman-compose|popd|pr|printcap|printenv|ps|pushd|pv|quota|quotacheck|quotactl|ram|rar|rcp|reboot|remsync|rename|renice|rev|rm|rmdir|rpm|rsync|scp|screen|sdiff|sed|sendmail|seq|service|sftp|sh|shellcheck|shuf|shutdown|sleep|slocate|sort|split|ssh|stat|strace|su|sudo|sum|suspend|swapon|sync|sysctl|tac|tail|tar|tee|time|timeout|top|touch|tr|traceroute|tsort|tty|umount|uname|unexpand|uniq|units|unrar|unshar|unzip|update-grub|uptime|useradd|userdel|usermod|users|uudecode|uuencode|v|vcpkg|vdir|vi|vim|virsh|vmstat|wait|watch|wc|wget|whereis|which|who|whoami|write|xargs|xdg-open|yarn|yes|zenity|zip|zsh|zypper)(?=$|[)\s;|&])/,lookbehind:!0},keyword:{pattern:/(^|[\s;|&]|[<>]\()(?:case|do|done|elif|else|esac|fi|for|function|if|in|select|then|until|while)(?=$|[)\s;|&])/,lookbehind:!0},builtin:{pattern:/(^|[\s;|&]|[<>]\()(?:\.|:|alias|bind|break|builtin|caller|cd|command|continue|declare|echo|enable|eval|exec|exit|export|getopts|hash|help|let|local|logout|mapfile|printf|pwd|read|readarray|readonly|return|set|shift|shopt|source|test|times|trap|type|typeset|ulimit|umask|unalias|unset)(?=$|[)\s;|&])/,lookbehind:!0,alias:"class-name"},boolean:{pattern:/(^|[\s;|&]|[<>]\()(?:false|true)(?=$|[)\s;|&])/,lookbehind:!0},"file-descriptor":{pattern:/\B&\d\b/,alias:"important"},operator:{pattern:/\d?<>|>\||\+=|=[=~]?|!=?|<<[<-]?|[&\d]?>>|\d[<>]&?|[<>][&=]?|&[>&]?|\|[&|]?/,inside:{"file-descriptor":{pattern:/^\d/,alias:"important"}}},punctuation:/\$?\(\(?|\)\)?|\.\.|[{}[\];\\]/,number:{pattern:/(^|\s)(?:[1-9]\d*|0)(?:[.,]\d+)?\b/,lookbehind:!0}},a.inside=e.languages.bash;for(var r=["comment","function-name","for-or-select","assign-left","parameter","string","environment","function","keyword","builtin","boolean","file-descriptor","operator","punctuation","number"],h=n.variable[1].inside,m=0;m<r.length;m++)h[r[m]]=e.languages.bash[r[m]];e.languages.sh=e.languages.bash,e.languages.shell=e.languages.bash})(d);var we=d;class Ue extends y{static observedAttributes=["count"];static states={count:je};connectedCallback(){this.set("parity",()=>this.get("count")%2?"odd":"even"),this.first(".increment").on("click",()=>this.set("count",(e)=>++e)),this.first(".decrement").on("click",()=>this.set("count",(e)=>--e)),this.first(".count").sync($("count")),this.first(".parity").sync($("parity"))}}Ue.define("my-counter");class ze extends y{connectedCallback(){this.first("span").sync($("name")),this.first("input").on("input",(e)=>this.set("name",e.target.value||void 0))}}ze.define("hello-world");class qe extends y{connectedCallback(){this.set("active",0);let e=this.querySelectorAll(".slide").length,t=(a,n)=>(a+n+e)%e;this.first(".prev").on("click",()=>this.set("active",(a)=>t(a,-1))),this.first(".next").on("click",()=>this.set("active",(a)=>t(a,1))),this.all(".slide").sync((a,n,r)=>M("active",()=>r===this.get("active"))(a,n)),this.all(".dots span").sync((a,n,r)=>M("active",()=>r===this.get("active"))(a,n))}}qe.define("my-slider");class Ze extends y{static observedAttributes=["collapsed"];static states={collapsed:I};connectedCallback(){let e=this.getAttribute("language")||"html",t=this.querySelector("code");if(t)this.set("code",t.textContent?.trim(),!1),B(()=>{let a=document.createElement("code");a.innerHTML=we.highlight(this.get("code"),we.languages[e],e),C(()=>{Array.from(a.childNodes).filter((n)=>n.nodeType!==Node.COMMENT_NODE).forEach((n)=>n.remove()),Array.from(a.childNodes).forEach((n)=>a.appendChild(n))},[a,"h"])}),this.first(".copy").on("click",async(a)=>{let n=a.target,r=n.textContent,h="success";try{await navigator.clipboard.writeText(t.textContent??"")}catch(m){console.error("Error when trying to use navigator.clipboard.writeText()",m),h="error"}n.set("disabled",!0),n.set("label",this.getAttribute(`copy-${h}`)),setTimeout(()=>{n.set("disabled",!1),n.set("label",r)},h==="success"?1000:3000)}),this.first(".overlay").on("click",()=>this.set("collapsed",!1)),this.self.sync(U("collapsed"))}}Ze.define("code-block");class Je extends y{static observedAttributes=["accordion"];static states={accordion:I};static consumedContexts=["media-viewport"];connectedCallback(){super.connectedCallback(),this.set("active",0,!1),setTimeout(()=>{if(this.get("media-viewport"))this.set("accordion",()=>["xs","sm"].includes(String(this.get("media-viewport"))))},0),this.self.sync(U("accordion")),this.first("menu").sync(F("ariaHidden","accordion")),this.all("menu button").on("click",(e,t)=>()=>this.set("active",t)).sync((e,t,a)=>F("ariaPressed",()=>this.get("active")===a)(e,t)),this.all("accordion-panel").pass({open:(e,t)=>()=>this.get("active")===t,collapsible:"accordion"})}}Je.define("tab-list");class Ye extends y{connectedCallback(){this.set("open",this.hasAttribute("open"),!1),this.set("collapsible",this.hasAttribute("collapsible"),!1),this.self.sync(U("open"),U("collapsible"),F("ariaHidden",()=>!this.get("open")&&!this.get("collapsible"))),this.first("details").sync(F("open"),F("ariaDisabled",()=>!this.get("collapsible")))}}Ye.define("accordion-panel");class Xe extends y{static observedAttributes=["disabled"];static states={disabled:I};connectedCallback(){this.first("button").sync($("label"),F("disabled"))}}Xe.define("input-button");var Ae=(e)=>typeof e==="number",oe=(e,t=!1)=>t?parseInt(e,10):parseFloat(e);class Be extends y{static observedAttributes=["value","description"];static states={value:(e)=>this.isNumber?e.map((t)=>oe(t,this.isInteger)).filter(Number.isFinite):e};connectedCallback(){this.input=this.querySelector("input"),this.isNumber=this.input&&this.input.type==="number",this.isInteger=this.hasAttribute("integer"),this.set("value",this.isNumber?this.input.valueAsNumber:this.input.value,!1),this.set("length",this.input.value.length),this.set("empty",()=>!this.get("length")),this.#t(),this.#a(),this.#n(),this.#r(),this.input.onchange=()=>this.#e(this.isNumber?this.input.valueAsNumber:this.input.value),this.input.oninput=()=>this.set("length",this.input.value.length),B(async()=>{let e=this.get("value"),t=this.getAttribute("validate");if(e&&t)await fetch(`${t}?name=${this.input.name}value=${this.input.value}`).then(async(a)=>{let n=await a.text();this.input.setCustomValidity(n),this.set("error",n)}).catch((a)=>this.set("error",a.message));if(this.isNumber&&!Ae(e))return this.set("value",oe(e,this.isInteger));if(this.isNumber&&!Number.isNaN(e))this.input.value=e})}clear(){this.input.value="",this.set("value",""),this.set("length",0),this.input.focus()}#e=(e)=>{if(this.set("value",e),this.set("error",this.input.validationMessage),typeof e==="function")e=this.get("value");if(this.input.value!==String(e))this.dispatchEvent(new CustomEvent("value-change",{detail:e,bubbles:!0}))};#t(){let e=this.first(".error");this.set("ariaInvalid",()=>String(Boolean(this.get("error")))),this.set("aria-errormessage",()=>this.get("error")?this.querySelector(".error")?.id:void 0),this.first(".error").sync($("error")),this.first("input").sync(F("ariaInvalid"),W("aria-errormessage"))}#a(){let e=this.first(".description");if(!e[0])return;let t=this.first("input"),a=this.input.maxLength,n=a&&e[0].target.dataset.remaining,r=e[0].target.textContent;this.set("description",n?()=>{let h=this.get("length");return h>0?n.replace("${x}",a-h):r}:r),this.set("aria-describedby",()=>this.get("description")?e[0].target.id:void 0),e.sync($("description")),t.sync(W("aria-describedby"))}#n(){let e=this.querySelector(".spinbutton");if(!this.isNumber||!e)return;let t=(f)=>maybe(oe(this.input[f],this.isInteger)).filter(Number.isFinite)[0],a=oe(e.dataset.step,this.isInteger),[n,r,h]=Number.isFinite(a)?[a,t("min"),t("max")]:[],m=(f)=>{let b=Math.round((h-r)/n),A=Math.round((f-r)*b/(h-r))/b;A=Math.min(Math.max(A,0),1);let S=A*(h-r)+r;return this.isInteger?Math.round(S):S};this.stepDown=(f=n)=>this.#e((b)=>m(b-f)),this.stepUp=(f=n)=>this.#e((b)=>m(b+f)),this.set("decrement-disabled",()=>Ae(r)&&this.get("value")-n<r),this.set("increment-disabled",()=>Ae(h)&&this.get("value")+n>h),this.first(".decrement").on("click",(f)=>this.stepDown(f.shiftKey?n*10:n)).sync(F("disabled","decrement-disabled")),this.first(".increment").on("click",(f)=>this.stepUp(f.shiftKey?n*10:n)).sync(F("disabled","increment-disabled")),this.input.onkeydown=(f)=>{if(["ArrowUp","ArrowDown"].includes(f.key)){if(f.stopPropagation(),f.preventDefault(),f.key==="ArrowDown")this.stepDown(f.shiftKey?n*10:n);if(f.key==="ArrowUp")this.stepUp(f.shiftKey?n*10:n)}}}#r(){this.first(".clear").on("click",()=>{this.clear(),this.input.focus()}).sync(M("hidden","empty"))}}Be.define("input-field");class Ie extends y{static observedAttributes=["checked"];static states={checked:I};connectedCallback(){this.first("input").on("change",(e)=>this.set("checked",Boolean(e.target.checked))).sync(F("checked")),this.self.sync(U("checked"))}}Ie.define("input-checkbox");class Qe extends y{static observedAttributes=["value"];connectedCallback(){this.self.sync(W("value")),this.all("input").on("change",(e)=>this.set("value",e.target.value)),this.all("label").sync((e,t)=>M("selected",()=>e.get("value")===t.querySelector("input")?.value)(e,t))}}Qe.define("input-radiogroup");class We extends y{static observedAttributes=["src"];static states={src:(e)=>e.map((t)=>{let a="";try{if(a=new URL(t,location.href),a.origin!==location.origin)throw new TypeError("Invalid URL origin")}catch(n){console.error(n,a),a=""}return a.toString()})};connectedCallback(){this.first(".loading").sync(F("ariaHidden",()=>!!this.get("error"))),this.first(".error").sync($("error")),F("ariaHidden",()=>!this.get("error")),B(()=>{let e=this.get("src");if(!e)return;fetch(e).then(async(t)=>{if(t.ok){let a=await t.text();C(()=>{this.root.innerHTML=a,this.root.querySelectorAll("script").forEach((n)=>{let r=document.createElement("script");r.appendChild(document.createTextNode(n.textContent??"")),this.root.appendChild(r),n.remove()},[this.root,"h"])}),this.set("error","")}else this.set("error",t.status+":"+t.statusText)}).catch((t)=>this.set("error",t))})}}We.define("lazy-load");var xe="media-motion",ue="media-theme",z="media-viewport",He="media-orientation";class Me extends y{static providedContexts=[xe,ue,z,He];connectedCallback(){super.connectedCallback();let e="light",t="dark",a="xs",n="sm",r="md",h="lg",m="xl",f="landscape",b="portrait",S=(()=>{let c=(j)=>{let T=this.getAttribute(j)?.trim();if(!T)return null;let K=T.match(/em$/)?"em":"px",V=maybe(parseFloat(T)).filter(Number.isFinite)[0];return V?V+K:null},g=c(n)||"32em",p=c(r)||"48em",k=c(h)||"72em",v=c(m)||"108em";return{sm:g,md:p,lg:k,xl:v}})(),N=matchMedia("(prefers-reduced-motion: reduce)"),G=matchMedia("(prefers-color-scheme: dark)"),q=matchMedia(`(min-width: ${S.sm})`),i=matchMedia(`(min-width: ${S.md})`),s=matchMedia(`(min-width: ${S.lg})`),u=matchMedia(`(min-width: ${S.xl})`),l=matchMedia("(orientation: landscape)"),o=()=>{if(u.matches)return m;if(s.matches)return h;if(i.matches)return r;if(q.matches)return n;return a};this.set(xe,N.matches),this.set(ue,G.matches?t:e),this.set(z,o()),this.set(He,l.matches?f:b),N.addEventListener("change",(c)=>this.set(xe,c.matches)),G.addEventListener("change",(c)=>this.set(ue,c.matches?t:e)),q.addEventListener("change",()=>this.set(z,o())),i.addEventListener("change",()=>this.set(z,o())),s.addEventListener("change",()=>this.set(z,o())),u.addEventListener("change",()=>this.set(z,o())),l.addEventListener("change",(c)=>{this.set(ue,c.matches?f:b)})}}Me.define("media-context");class Ke extends y{static states={tasks:[],filter:"all"};connectedCallback(){this.#e(),this.self.on("click",(e)=>{if(e.target.localName==="button")this.removeItem(e.target)}).sync(W("filter")),this.set("count",()=>{let e=this.get("tasks").map((n)=>n.signals.get("checked")),t=e.filter((n)=>n?.get()).length,a=e.length;return{active:a-t,completed:t,total:a}})}addItem=(e)=>{let t=this.querySelector("template")?.content.cloneNode(!0);if(t&&t.querySelector("span"))t.querySelector("span").textContent=e,this.querySelector("ol")?.appendChild(t),this.#e()};removeItem=(e)=>{e.closest("li")?.remove(),this.#e()};clearCompleted=()=>{this.get("tasks").filter((e)=>e.get("checked")).forEach((e)=>e.parentElement?.remove()),this.#e()};#e(){this.set("tasks",Array.from(this.querySelectorAll("input-checkbox")))}}Ke.define("todo-list");class Ve extends y{connectedCallback(){let e=this.querySelector("todo-list"),t=this.querySelector("input-radiogroup");this.self.on("add-todo",(a)=>e?.addItem(a.detail)),this.first("todo-count").pass({active:()=>(e?.get("count")).active}),this.first("todo-list").pass({filter:()=>t?.get("value")}),this.first(".clear-completed").on("click",()=>e?.clearCompleted()).pass({disabled:()=>!(e?.get("count")).completed})}}Ve.define("todo-app");class et extends y{static states={active:0};connectedCallback(){this.first(".count").sync($("active")),this.first(".singular").sync(F("ariaHidden",()=>this.get("active")>1)),this.first(".plural").sync(F("ariaHidden",()=>this.get("active")===1)),this.first(".remaining").sync(F("ariaHidden",()=>!this.get("active"))),this.first(".all-done").sync(F("ariaHidden",()=>!!this.get("active")))}}et.define("todo-count");class tt extends y{connectedCallback(){let e=this.querySelector("input-field");this.first("form").on("submit",(t)=>{t.preventDefault(),setTimeout(()=>{this.dispatchEvent(new CustomEvent("add-todo",{bubbles:!0,detail:e?.get("value")})),e?.clear()},0)}),this.first("input-button").pass({disabled:()=>e?.get("empty")})}}tt.define("todo-form");export{Qe as InputRadiogroup};
+var bu=Object.create;var{getPrototypeOf:Gu,defineProperty:Of,getOwnPropertyNames:Hu}=Object;var Cu=Object.prototype.hasOwnProperty;var Ku=(f,u,w)=>{w=f!=null?bu(Gu(f)):{};let J=u||!f||!f.__esModule?Of(w,"default",{value:f,enumerable:!0}):w;for(let h of Hu(f))if(!Cu.call(J,h))Of(J,h,{get:()=>f[h],enumerable:!0});return J};var zu=(f,u)=>()=>(u||f((u={exports:{}}).exports,u),u.exports);var rf=zu((F0,Qf)=>{var iu=typeof window!=="undefined"?window:typeof WorkerGlobalScope!=="undefined"&&self instanceof WorkerGlobalScope?self:{},U=function(f){var u=/(?:^|\s)lang(?:uage)?-([\w-]+)(?=\s|$)/i,w=0,J={},h={manual:f.Prism&&f.Prism.manual,disableWorkerMessageHandler:f.Prism&&f.Prism.disableWorkerMessageHandler,util:{encode:function Z(q){if(q instanceof X)return new X(q.type,Z(q.content),q.alias);else if(Array.isArray(q))return q.map(Z);else return q.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/\u00a0/g," ")},type:function(Z){return Object.prototype.toString.call(Z).slice(8,-1)},objId:function(Z){if(!Z.__id)Object.defineProperty(Z,"__id",{value:++w});return Z.__id},clone:function Z(q,Q){Q=Q||{};var Y,F;switch(h.util.type(q)){case"Object":if(F=h.util.objId(q),Q[F])return Q[F];Y={},Q[F]=Y;for(var b in q)if(q.hasOwnProperty(b))Y[b]=Z(q[b],Q);return Y;case"Array":if(F=h.util.objId(q),Q[F])return Q[F];return Y=[],Q[F]=Y,q.forEach(function(C,W){Y[W]=Z(C,Q)}),Y;default:return q}},getLanguage:function(Z){while(Z){var q=u.exec(Z.className);if(q)return q[1].toLowerCase();Z=Z.parentElement}return"none"},setLanguage:function(Z,q){Z.className=Z.className.replace(RegExp(u,"gi"),""),Z.classList.add("language-"+q)},currentScript:function(){if(typeof document==="undefined")return null;if("currentScript"in document)return document.currentScript;try{throw new Error}catch(Y){var Z=(/at [^(\r\n]*\((.*):[^:]+:[^:]+\)$/i.exec(Y.stack)||[])[1];if(Z){var q=document.getElementsByTagName("script");for(var Q in q)if(q[Q].src==Z)return q[Q]}return null}},isActive:function(Z,q,Q){var Y="no-"+q;while(Z){var F=Z.classList;if(F.contains(q))return!0;if(F.contains(Y))return!1;Z=Z.parentElement}return!!Q}},languages:{plain:J,plaintext:J,text:J,txt:J,extend:function(Z,q){var Q=h.util.clone(h.languages[Z]);for(var Y in q)Q[Y]=q[Y];return Q},insertBefore:function(Z,q,Q,Y){Y=Y||h.languages;var F=Y[Z],b={};for(var C in F)if(F.hasOwnProperty(C)){if(C==q){for(var W in Q)if(Q.hasOwnProperty(W))b[W]=Q[W]}if(!Q.hasOwnProperty(C))b[C]=F[C]}var y=Y[Z];return Y[Z]=b,h.languages.DFS(h.languages,function(E,T){if(T===y&&E!=Z)this[E]=b}),b},DFS:function Z(q,Q,Y,F){F=F||{};var b=h.util.objId;for(var C in q)if(q.hasOwnProperty(C)){Q.call(q,C,q[C],Y||C);var W=q[C],y=h.util.type(W);if(y==="Object"&&!F[b(W)])F[b(W)]=!0,Z(W,Q,null,F);else if(y==="Array"&&!F[b(W)])F[b(W)]=!0,Z(W,Q,C,F)}}},plugins:{},highlightAll:function(Z,q){h.highlightAllUnder(document,Z,q)},highlightAllUnder:function(Z,q,Q){var Y={callback:Q,container:Z,selector:'code[class*="language-"], [class*="language-"] code, code[class*="lang-"], [class*="lang-"] code'};h.hooks.run("before-highlightall",Y),Y.elements=Array.prototype.slice.apply(Y.container.querySelectorAll(Y.selector)),h.hooks.run("before-all-elements-highlight",Y);for(var F=0,b;b=Y.elements[F++];)h.highlightElement(b,q===!0,Y.callback)},highlightElement:function(Z,q,Q){var Y=h.util.getLanguage(Z),F=h.languages[Y];h.util.setLanguage(Z,Y);var b=Z.parentElement;if(b&&b.nodeName.toLowerCase()==="pre")h.util.setLanguage(b,Y);var C=Z.textContent,W={element:Z,language:Y,grammar:F,code:C};function y(T){W.highlightedCode=T,h.hooks.run("before-insert",W),W.element.innerHTML=W.highlightedCode,h.hooks.run("after-highlight",W),h.hooks.run("complete",W),Q&&Q.call(W.element)}if(h.hooks.run("before-sanity-check",W),b=W.element.parentElement,b&&b.nodeName.toLowerCase()==="pre"&&!b.hasAttribute("tabindex"))b.setAttribute("tabindex","0");if(!W.code){h.hooks.run("complete",W),Q&&Q.call(W.element);return}if(h.hooks.run("before-highlight",W),!W.grammar){y(h.util.encode(W.code));return}if(q&&f.Worker){var E=new Worker(h.filename);E.onmessage=function(T){y(T.data)},E.postMessage(JSON.stringify({language:W.language,code:W.code,immediateClose:!0}))}else y(h.highlight(W.code,W.grammar,W.language))},highlight:function(Z,q,Q){var Y={code:Z,grammar:q,language:Q};if(h.hooks.run("before-tokenize",Y),!Y.grammar)throw new Error('The language "'+Y.language+'" has no grammar.');return Y.tokens=h.tokenize(Y.code,Y.grammar),h.hooks.run("after-tokenize",Y),X.stringify(h.util.encode(Y.tokens),Y.language)},tokenize:function(Z,q){var Q=q.rest;if(Q){for(var Y in Q)q[Y]=Q[Y];delete q.rest}var F=new z;return L(F,F.head,Z),$(Z,F,q,F.head,0),v(F)},hooks:{all:{},add:function(Z,q){var Q=h.hooks.all;Q[Z]=Q[Z]||[],Q[Z].push(q)},run:function(Z,q){var Q=h.hooks.all[Z];if(!Q||!Q.length)return;for(var Y=0,F;F=Q[Y++];)F(q)}},Token:X};f.Prism=h;function X(Z,q,Q,Y){this.type=Z,this.content=q,this.alias=Q,this.length=(Y||"").length|0}X.stringify=function Z(q,Q){if(typeof q=="string")return q;if(Array.isArray(q)){var Y="";return q.forEach(function(y){Y+=Z(y,Q)}),Y}var F={type:q.type,content:Z(q.content,Q),tag:"span",classes:["token",q.type],attributes:{},language:Q},b=q.alias;if(b)if(Array.isArray(b))Array.prototype.push.apply(F.classes,b);else F.classes.push(b);h.hooks.run("wrap",F);var C="";for(var W in F.attributes)C+=" "+W+'="'+(F.attributes[W]||"").replace(/"/g,"&quot;")+'"';return"<"+F.tag+' class="'+F.classes.join(" ")+'"'+C+">"+F.content+"</"+F.tag+">"};function G(Z,q,Q,Y){Z.lastIndex=q;var F=Z.exec(Q);if(F&&Y&&F[1]){var b=F[1].length;F.index+=b,F[0]=F[0].slice(b)}return F}function $(Z,q,Q,Y,F,b){for(var C in Q){if(!Q.hasOwnProperty(C)||!Q[C])continue;var W=Q[C];W=Array.isArray(W)?W:[W];for(var y=0;y<W.length;++y){if(b&&b.cause==C+","+y)return;var E=W[y],T=E.inside,Ef=!!E.lookbehind,Mf=!!E.greedy,Fu=E.alias;if(Mf&&!E.pattern.global){var Uu=E.pattern.toString().match(/[imsuy]*$/)[0];E.pattern=RegExp(E.pattern.source,Uu+"g")}var Rf=E.pattern||E;for(var M=Y.next,I=F;M!==q.tail;I+=M.value.length,M=M.next){if(b&&I>=b.reach)break;var d=M.value;if(q.length>Z.length)return;if(d instanceof X)continue;var e=1,x;if(Mf){if(x=G(Rf,I,Z,Ef),!x||x.index>=Z.length)break;var ff=x.index,Xu=x.index+x[0].length,g=I;g+=M.value.length;while(ff>=g)M=M.next,g+=M.value.length;if(g-=M.value.length,I=g,M.value instanceof X)continue;for(var a=M;a!==q.tail&&(g<Xu||typeof a.value==="string");a=a.next)e++,g+=a.value.length;e--,d=Z.slice(I,g),x.index-=I}else if(x=G(Rf,0,d,Ef),!x)continue;var ff=x.index,uf=x[0],Uf=d.slice(0,ff),Sf=d.slice(ff+uf.length),Xf=I+d.length;if(b&&Xf>b.reach)b.reach=Xf;var wf=M.prev;if(Uf)wf=L(q,wf,Uf),I+=Uf.length;j(q,wf,e);var Wu=new X(C,T?h.tokenize(uf,T):uf,Fu,uf);if(M=L(q,wf,Wu),Sf)L(q,M,Sf);if(e>1){var Wf={cause:C+","+y,reach:Xf};if($(Z,q,Q,M.prev,I,Wf),b&&Wf.reach>b.reach)b.reach=Wf.reach}}}}}function z(){var Z={value:null,prev:null,next:null},q={value:null,prev:Z,next:null};Z.next=q,this.head=Z,this.tail=q,this.length=0}function L(Z,q,Q){var Y=q.next,F={value:Q,prev:q,next:Y};return q.next=F,Y.prev=F,Z.length++,F}function j(Z,q,Q){var Y=q.next;for(var F=0;F<Q&&Y!==Z.tail;F++)Y=Y.next;q.next=Y,Y.prev=q,Z.length-=F}function v(Z){var q=[],Q=Z.head.next;while(Q!==Z.tail)q.push(Q.value),Q=Q.next;return q}if(!f.document){if(!f.addEventListener)return h;if(!h.disableWorkerMessageHandler)f.addEventListener("message",function(Z){var q=JSON.parse(Z.data),Q=q.language,Y=q.code,F=q.immediateClose;if(f.postMessage(h.highlight(Y,h.languages[Q],Q)),F)f.close()},!1);return h}var B=h.util.currentScript();if(B){if(h.filename=B.src,B.hasAttribute("data-manual"))h.manual=!0}function K(){if(!h.manual)h.highlightAll()}if(!h.manual){var V=document.readyState;if(V==="loading"||V==="interactive"&&B&&B.defer)document.addEventListener("DOMContentLoaded",K);else if(window.requestAnimationFrame)window.requestAnimationFrame(K);else window.setTimeout(K,16)}return h}(iu);if(typeof Qf!=="undefined"&&Qf.exports)Qf.exports=U;if(typeof global!=="undefined")global.Prism=U;U.languages.markup={comment:{pattern:/<!--(?:(?!<!--)[\s\S])*?-->/,greedy:!0},prolog:{pattern:/<\?[\s\S]+?\?>/,greedy:!0},doctype:{pattern:/<!DOCTYPE(?:[^>"'[\]]|"[^"]*"|'[^']*')+(?:\[(?:[^<"'\]]|"[^"]*"|'[^']*'|<(?!!--)|<!--(?:[^-]|-(?!->))*-->)*\]\s*)?>/i,greedy:!0,inside:{"internal-subset":{pattern:/(^[^\[]*\[)[\s\S]+(?=\]>$)/,lookbehind:!0,greedy:!0,inside:null},string:{pattern:/"[^"]*"|'[^']*'/,greedy:!0},punctuation:/^<!|>$|[[\]]/,"doctype-tag":/^DOCTYPE/i,name:/[^\s<>'"]+/}},cdata:{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,greedy:!0},tag:{pattern:/<\/?(?!\d)[^\s>\/=$<%]+(?:\s(?:\s*[^\s>\/=]+(?:\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+(?=[\s>]))|(?=[\s/>])))+)?\s*\/?>/,greedy:!0,inside:{tag:{pattern:/^<\/?[^\s>\/]+/,inside:{punctuation:/^<\/?/,namespace:/^[^\s>\/:]+:/}},"special-attr":[],"attr-value":{pattern:/=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+)/,inside:{punctuation:[{pattern:/^=/,alias:"attr-equals"},{pattern:/^(\s*)["']|["']$/,lookbehind:!0}]}},punctuation:/\/?>/,"attr-name":{pattern:/[^\s>\/]+/,inside:{namespace:/^[^\s>\/:]+:/}}}},entity:[{pattern:/&[\da-z]{1,8};/i,alias:"named-entity"},/&#x?[\da-f]{1,8};/i]};U.languages.markup.tag.inside["attr-value"].inside.entity=U.languages.markup.entity;U.languages.markup.doctype.inside["internal-subset"].inside=U.languages.markup;U.hooks.add("wrap",function(f){if(f.type==="entity")f.attributes.title=f.content.replace(/&amp;/,"&")});Object.defineProperty(U.languages.markup.tag,"addInlined",{value:function f(u,w){var J={};J["language-"+w]={pattern:/(^<!\[CDATA\[)[\s\S]+?(?=\]\]>$)/i,lookbehind:!0,inside:U.languages[w]},J.cdata=/^<!\[CDATA\[|\]\]>$/i;var h={"included-cdata":{pattern:/<!\[CDATA\[[\s\S]*?\]\]>/i,inside:J}};h["language-"+w]={pattern:/[\s\S]+/,inside:U.languages[w]};var X={};X[u]={pattern:RegExp(/(<__[^>]*>)(?:<!\[CDATA\[(?:[^\]]|\](?!\]>))*\]\]>|(?!<!\[CDATA\[)[\s\S])*?(?=<\/__>)/.source.replace(/__/g,function(){return u}),"i"),lookbehind:!0,greedy:!0,inside:h},U.languages.insertBefore("markup","cdata",X)}});Object.defineProperty(U.languages.markup.tag,"addAttribute",{value:function(f,u){U.languages.markup.tag.inside["special-attr"].push({pattern:RegExp(/(^|["'\s])/.source+"(?:"+f+")"+/\s*=\s*(?:"[^"]*"|'[^']*'|[^\s'">=]+(?=[\s>]))/.source,"i"),lookbehind:!0,inside:{"attr-name":/^[^\s=]+/,"attr-value":{pattern:/=[\s\S]+/,inside:{value:{pattern:/(^=\s*(["']|(?!["'])))\S[\s\S]*(?=\2$)/,lookbehind:!0,alias:[u,"language-"+u],inside:U.languages[u]},punctuation:[{pattern:/^=/,alias:"attr-equals"},/"|'/]}}}})}});U.languages.html=U.languages.markup;U.languages.mathml=U.languages.markup;U.languages.svg=U.languages.markup;U.languages.xml=U.languages.extend("markup",{});U.languages.ssml=U.languages.xml;U.languages.atom=U.languages.xml;U.languages.rss=U.languages.xml;(function(f){var u=/(?:"(?:\\(?:\r\n|[\s\S])|[^"\\\r\n])*"|'(?:\\(?:\r\n|[\s\S])|[^'\\\r\n])*')/;f.languages.css={comment:/\/\*[\s\S]*?\*\//,atrule:{pattern:RegExp("@[\\w-](?:"+/[^;{\s"']|\s+(?!\s)/.source+"|"+u.source+")*?"+/(?:;|(?=\s*\{))/.source),inside:{rule:/^@[\w-]+/,"selector-function-argument":{pattern:/(\bselector\s*\(\s*(?![\s)]))(?:[^()\s]|\s+(?![\s)])|\((?:[^()]|\([^()]*\))*\))+(?=\s*\))/,lookbehind:!0,alias:"selector"},keyword:{pattern:/(^|[^\w-])(?:and|not|only|or)(?![\w-])/,lookbehind:!0}}},url:{pattern:RegExp("\\burl\\((?:"+u.source+"|"+/(?:[^\\\r\n()"']|\\[\s\S])*/.source+")\\)","i"),greedy:!0,inside:{function:/^url/i,punctuation:/^\(|\)$/,string:{pattern:RegExp("^"+u.source+"$"),alias:"url"}}},selector:{pattern:RegExp('(^|[{}\\s])[^{}\\s](?:[^{};"\'\\s]|\\s+(?![\\s{])|'+u.source+")*(?=\\s*\\{)"),lookbehind:!0},string:{pattern:u,greedy:!0},property:{pattern:/(^|[^-\w\xA0-\uFFFF])(?!\s)[-_a-z\xA0-\uFFFF](?:(?!\s)[-\w\xA0-\uFFFF])*(?=\s*:)/i,lookbehind:!0},important:/!important\b/i,function:{pattern:/(^|[^-a-z0-9])[-a-z0-9]+(?=\()/i,lookbehind:!0},punctuation:/[(){};:,]/},f.languages.css.atrule.inside.rest=f.languages.css;var w=f.languages.markup;if(w)w.tag.addInlined("style","css"),w.tag.addAttribute("style","css")})(U);U.languages.clike={comment:[{pattern:/(^|[^\\])\/\*[\s\S]*?(?:\*\/|$)/,lookbehind:!0,greedy:!0},{pattern:/(^|[^\\:])\/\/.*/,lookbehind:!0,greedy:!0}],string:{pattern:/(["'])(?:\\(?:\r\n|[\s\S])|(?!\1)[^\\\r\n])*\1/,greedy:!0},"class-name":{pattern:/(\b(?:class|extends|implements|instanceof|interface|new|trait)\s+|\bcatch\s+\()[\w.\\]+/i,lookbehind:!0,inside:{punctuation:/[.\\]/}},keyword:/\b(?:break|catch|continue|do|else|finally|for|function|if|in|instanceof|new|null|return|throw|try|while)\b/,boolean:/\b(?:false|true)\b/,function:/\b\w+(?=\()/,number:/\b0x[\da-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:e[+-]?\d+)?/i,operator:/[<>]=?|[!=]=?=?|--?|\+\+?|&&?|\|\|?|[?*/~^%]/,punctuation:/[{}[\];(),.:]/};U.languages.javascript=U.languages.extend("clike",{"class-name":[U.languages.clike["class-name"],{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$A-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\.(?:constructor|prototype))/,lookbehind:!0}],keyword:[{pattern:/((?:^|\})\s*)catch\b/,lookbehind:!0},{pattern:/(^|[^.]|\.\.\.\s*)\b(?:as|assert(?=\s*\{)|async(?=\s*(?:function\b|\(|[$\w\xA0-\uFFFF]|$))|await|break|case|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally(?=\s*(?:\{|$))|for|from(?=\s*(?:['"]|$))|function|(?:get|set)(?=\s*(?:[#\[$\w\xA0-\uFFFF]|$))|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)\b/,lookbehind:!0}],function:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*(?:\.\s*(?:apply|bind|call)\s*)?\()/,number:{pattern:RegExp(/(^|[^\w$])/.source+"(?:"+(/NaN|Infinity/.source+"|"+/0[bB][01]+(?:_[01]+)*n?/.source+"|"+/0[oO][0-7]+(?:_[0-7]+)*n?/.source+"|"+/0[xX][\dA-Fa-f]+(?:_[\dA-Fa-f]+)*n?/.source+"|"+/\d+(?:_\d+)*n/.source+"|"+/(?:\d+(?:_\d+)*(?:\.(?:\d+(?:_\d+)*)?)?|\.\d+(?:_\d+)*)(?:[Ee][+-]?\d+(?:_\d+)*)?/.source)+")"+/(?![\w$])/.source),lookbehind:!0},operator:/--|\+\+|\*\*=?|=>|&&=?|\|\|=?|[!=]==|<<=?|>>>?=?|[-+*/%&|^!=<>]=?|\.{3}|\?\?=?|\?\.?|[~:]/});U.languages.javascript["class-name"][0].pattern=/(\b(?:class|extends|implements|instanceof|interface|new)\s+)[\w.\\]+/;U.languages.insertBefore("javascript","keyword",{regex:{pattern:RegExp(/((?:^|[^$\w\xA0-\uFFFF."'\])\s]|\b(?:return|yield))\s*)/.source+/\//.source+"(?:"+/(?:\[(?:[^\]\\\r\n]|\\.)*\]|\\.|[^/\\\[\r\n])+\/[dgimyus]{0,7}/.source+"|"+/(?:\[(?:[^[\]\\\r\n]|\\.|\[(?:[^[\]\\\r\n]|\\.|\[(?:[^[\]\\\r\n]|\\.)*\])*\])*\]|\\.|[^/\\\[\r\n])+\/[dgimyus]{0,7}v[dgimyus]{0,7}/.source+")"+/(?=(?:\s|\/\*(?:[^*]|\*(?!\/))*\*\/)*(?:$|[\r\n,.;:})\]]|\/\/))/.source),lookbehind:!0,greedy:!0,inside:{"regex-source":{pattern:/^(\/)[\s\S]+(?=\/[a-z]*$)/,lookbehind:!0,alias:"language-regex",inside:U.languages.regex},"regex-delimiter":/^\/|\/$/,"regex-flags":/^[a-z]+$/}},"function-variable":{pattern:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*[=:]\s*(?:async\s*)?(?:\bfunction\b|(?:\((?:[^()]|\([^()]*\))*\)|(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)\s*=>))/,alias:"function"},parameter:[{pattern:/(function(?:\s+(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*)?\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\))/,lookbehind:!0,inside:U.languages.javascript},{pattern:/(^|[^$\w\xA0-\uFFFF])(?!\s)[_$a-z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*=>)/i,lookbehind:!0,inside:U.languages.javascript},{pattern:/(\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*=>)/,lookbehind:!0,inside:U.languages.javascript},{pattern:/((?:\b|\s|^)(?!(?:as|async|await|break|case|catch|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally|for|from|function|get|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|set|static|super|switch|this|throw|try|typeof|undefined|var|void|while|with|yield)(?![$\w\xA0-\uFFFF]))(?:(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*\s*)\(\s*|\]\s*\(\s*)(?!\s)(?:[^()\s]|\s+(?![\s)])|\([^()]*\))+(?=\s*\)\s*\{)/,lookbehind:!0,inside:U.languages.javascript}],constant:/\b[A-Z](?:[A-Z_]|\dx?)*\b/});U.languages.insertBefore("javascript","string",{hashbang:{pattern:/^#!.*/,greedy:!0,alias:"comment"},"template-string":{pattern:/`(?:\\[\s\S]|\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}|(?!\$\{)[^\\`])*`/,greedy:!0,inside:{"template-punctuation":{pattern:/^`|`$/,alias:"string"},interpolation:{pattern:/((?:^|[^\\])(?:\\{2})*)\$\{(?:[^{}]|\{(?:[^{}]|\{[^}]*\})*\})+\}/,lookbehind:!0,inside:{"interpolation-punctuation":{pattern:/^\$\{|\}$/,alias:"punctuation"},rest:U.languages.javascript}},string:/[\s\S]+/}},"string-property":{pattern:/((?:^|[,{])[ \t]*)(["'])(?:\\(?:\r\n|[\s\S])|(?!\2)[^\\\r\n])*\2(?=\s*:)/m,lookbehind:!0,greedy:!0,alias:"property"}});U.languages.insertBefore("javascript","operator",{"literal-property":{pattern:/((?:^|[,{])[ \t]*)(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?=\s*:)/m,lookbehind:!0,alias:"property"}});if(U.languages.markup)U.languages.markup.tag.addInlined("script","javascript"),U.languages.markup.tag.addAttribute(/on(?:abort|blur|change|click|composition(?:end|start|update)|dblclick|error|focus(?:in|out)?|key(?:down|up)|load|mouse(?:down|enter|leave|move|out|over|up)|reset|resize|scroll|select|slotchange|submit|unload|wheel)/.source,"javascript");U.languages.js=U.languages.javascript;(function(){if(typeof U==="undefined"||typeof document==="undefined")return;if(!Element.prototype.matches)Element.prototype.matches=Element.prototype.msMatchesSelector||Element.prototype.webkitMatchesSelector;var f="Loading\u2026",u=function(B,K){return"\u2716 Error "+B+" while fetching file: "+K},w="\u2716 Error: File does not exist or is empty",J={js:"javascript",py:"python",rb:"ruby",ps1:"powershell",psm1:"powershell",sh:"bash",bat:"batch",h:"c",tex:"latex"},h="data-src-status",X="loading",G="loaded",$="failed",z="pre[data-src]:not(["+h+'="'+G+'"]):not(['+h+'="'+X+'"])';function L(B,K,V){var Z=new XMLHttpRequest;Z.open("GET",B,!0),Z.onreadystatechange=function(){if(Z.readyState==4)if(Z.status<400&&Z.responseText)K(Z.responseText);else if(Z.status>=400)V(u(Z.status,Z.statusText));else V(w)},Z.send(null)}function j(B){var K=/^\s*(\d+)\s*(?:(,)\s*(?:(\d+)\s*)?)?$/.exec(B||"");if(K){var V=Number(K[1]),Z=K[2],q=K[3];if(!Z)return[V,V];if(!q)return[V,void 0];return[V,Number(q)]}return}U.hooks.add("before-highlightall",function(B){B.selector+=", "+z}),U.hooks.add("before-sanity-check",function(B){var K=B.element;if(K.matches(z)){B.code="",K.setAttribute(h,X);var V=K.appendChild(document.createElement("CODE"));V.textContent=f;var Z=K.getAttribute("data-src"),q=B.language;if(q==="none"){var Q=(/\.(\w+)$/.exec(Z)||[,"none"])[1];q=J[Q]||Q}U.util.setLanguage(V,q),U.util.setLanguage(K,q);var Y=U.plugins.autoloader;if(Y)Y.loadLanguages(q);L(Z,function(F){K.setAttribute(h,G);var b=j(K.getAttribute("data-range"));if(b){var C=F.split(/\r\n?|\n/g),W=b[0],y=b[1]==null?C.length:b[1];if(W<0)W+=C.length;if(W=Math.max(0,Math.min(W-1,C.length)),y<0)y+=C.length;if(y=Math.max(0,Math.min(y,C.length)),F=C.slice(W,y).join("\n"),!K.hasAttribute("data-start"))K.setAttribute("data-start",String(W+1))}V.textContent=F,U.highlightElement(V)},function(F){K.setAttribute(h,$),V.textContent=F})}}),U.plugins.fileHighlight={highlight:function B(K){var V=(K||document).querySelectorAll(z);for(var Z=0,q;q=V[Z++];)U.highlightElement(q)}};var v=!1;U.fileHighlight=function(){if(!v)console.warn("Prism.fileHighlight is deprecated. Use `Prism.plugins.fileHighlight.highlight` instead."),v=!0;U.plugins.fileHighlight.highlight.apply(this,arguments)}})()});var zf=(f)=>typeof f==="function",yu=(f)=>zf(f)&&/^async\s+/.test(f.toString()),yf=(f)=>(u)=>u instanceof f,xf=yf(Error),$u=yf(Promise),Lf="Computed",$f=(f,u)=>{u=u??yu(f);let w=[],J,h=null,X=!0,G=()=>{if(X=!0,u)Af(w)},$={[Symbol.toStringTag]:Lf,get:()=>{if(u)Nf(w);if(!u||X)Tf(()=>{let z=(j)=>{J=j,X=!1,h=null},L=(j)=>{h=xf(j)?j:new Error(`Computed function failed: ${j}`)};try{let j=f(J);$u(j)?j.then(z).catch(L):z(j)}catch(j){L(j)}},G);if(xf(h))throw h;return J},map:(z)=>$f(()=>z($.get()))};return $},Bu=(f)=>!!f&&typeof f==="object"&&f[Symbol.toStringTag]===Lf,p,Vu=!1,ju=[],Gf=(f)=>Cf(f)||Bu(f),If=(f,u=!1)=>Gf(f)?f:zf(f)?$f(f,u):Du(f),Nf=(f)=>{if(p&&!f.includes(p))f.push(p)},Af=(f)=>f.forEach((u)=>Vu?ju.push(u):u()),Tf=(f,u)=>{let w=p;p=u,f(),p=w};var Hf=Symbol();class Bf{$;watchers=[];constructor(f){this.value=f}get(){return Nf(this.watchers),this.value}set(f){if(Hf!==f){let u=zf(f)?f(this.value):f;if(Object.is(this.value,u))return;this.value=u}if(Af(this.watchers),Hf===f)this.watchers=[]}map(f){return $f(()=>f(this.get()))}}var Du=(f)=>new Bf(f),Cf=yf(Bf),l=(f)=>{let u=()=>Tf(()=>{try{f()}catch(w){console.error(w)}},u);u()};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=(f)=>setTimeout(f,16);var Jf=new Map,m=[],hf,gf=()=>{hf=null,m.forEach((f)=>f()),m=[],Jf.clear()},Eu=()=>{if(hf)cancelAnimationFrame(hf);hf=requestAnimationFrame(gf)};queueMicrotask(gf);var N=(f,u)=>new Promise((w,J)=>{let h=()=>{try{w(f())}catch(X){J(X)}};if(u){let[X,G]=u;if(!Jf.has(X))Jf.set(X,new Map);let $=Jf.get(X);if($.has(G)){let z=m.indexOf(f);if(z>-1)m.splice(z,1)}$.set(G,h)}m.push(h),Eu()});var Mu=(f)=>f.nodeType===Node.COMMENT_NODE,Ru=(f)=>!/^on/i.test(f),Su=(f)=>{if(/^(mailto|tel):/i.test(f))return!0;if(f.includes("://"))try{let u=new URL(f,window.location.origin);return!["http:","https:","ftp:"].includes(u.protocol)}catch(u){return!0}return!0},Ou=(f,u,w)=>{if(!Ru(u))throw new Error(`Unsafe attribute: ${u}`);if(w=String(w).trim(),!Su(w))throw new Error(`Unsafe URL for ${u}: ${w}`);f.setAttribute(u,w)};var xu=(f)=>N(()=>{return f.remove(),null},[f,"r"]),Lu=(f,u)=>N(()=>{return Array.from(f.childNodes).filter((w)=>!Mu(w)).forEach((w)=>w.remove()),f.append(document.createTextNode(u)),f},[f,"t"]),Iu=(f,u,w)=>N(()=>{return Ou(f,u,w),f},[f,`a:${u}`]),Nu=(f,u)=>N(()=>{return f.removeAttribute(u),f},[f,`a:${u}`]),Au=(f,u,w)=>N(()=>{return f.toggleAttribute(u,w),f},[f,`a:${u}`]),Tu=(f,u,w)=>N(()=>{return f.classList.toggle(u,w),f},[f,`c:${u}`]);var c=(f)=>typeof f==="function",_f=(f)=>!!f&&typeof f==="object",gu=(f)=>typeof f==="number",Zf=(f)=>typeof f==="string",_u=(f)=>typeof f==="symbol",vu=(f)=>Zf(f)||_u(f)||gu(f),S=!1,Pu="debug";var cu="warn",n="error",lu=(f)=>f?`#${f}`:"",su=(f)=>f.length?`.${Array.from(f).join(".")}`:"",o=(f)=>`<${f.localName}${lu(f.id)}${su(f.classList)}>`,P=(f)=>Zf(f)?`"${f}"`:_f(f)?JSON.stringify(f):String(f),O=(f,u,w=Pu)=>{if(S||[n,cu].includes(w))console[w](u,f);return f},ku=(f)=>c(f)&&f.length===2,bf=(f,u,w=0)=>ku(f)?f(u,w):f;class qf{$;A;constructor(f,u=[f]){this.host=f,this.targets=u}on(f,u){return this.targets.forEach((w,J)=>w.addEventListener(f,bf(u,w,J))),this}off(f,u){return this.targets.forEach((w,J)=>w.removeEventListener(f,bf(u,w,J))),this}pass(f){return this.targets.forEach(async(u,w)=>{if(await H.registry.whenDefined(u.localName),u instanceof H)Object.entries(f).forEach(([J,h])=>{let X=bf(h,u,w),G=vu(X)?this.host.signals.get(X):If(X,!0);if(G)u.set(J,G);else O(h,`Invalid source for state ${P(J)}`,n)});else O(u,"Target is not a UIElement",n)}),this}sync(...f){return this.targets.forEach((u,w)=>f.forEach((J)=>J(this.host,u,w))),this}}var du=(f,u)=>{if(u==null)return;let w=f(u);return Number.isFinite(w)?w:void 0},Kf=(f,u,w,J=void 0)=>{let h=f.constructor.states[u];return c(h)&&!!h.length?h(w,f,J):w},_=(f)=>f!=null,vf=(f)=>du(parseInt,f);var Pf="context-request";class cf extends Event{$;A;B;constructor(f,u,w=!1){super(Pf,{bubbles:!0,composed:!0});this.context=f,this.callback=u,this.subscribe=w}}var pu=(f)=>{let u=f.constructor,w=u.consumedContexts||[];setTimeout(()=>{for(let h of w)f.dispatchEvent(new cf(h,(X)=>f.set(String(h),X)))});let J=u.providedContexts||[];if(!J.length)return!1;return f.addEventListener(Pf,(h)=>{let{context:X,callback:G}=h;if(!J.includes(X)||!c(G))return;h.stopPropagation(),G(f.signals.get(String(X)))}),!0};class H extends HTMLElement{static registry=customElements;static states={};static observedAttributes;static consumedContexts;static providedContexts;static define(f){try{if(H.registry.define(f,this),S)O(f,"Registered custom element")}catch(u){O(u,`Failed to register custom element ${f}`,n)}}signals=new Map;self=new qf(this);root=this.shadowRoot||this;debug=!1;attributeChangedCallback(f,u,w){if(w===u)return;if(S&&this.debug)O(`${P(u)} => ${P(w)}`,`Attribute "${f}" of ${o(this)} changed`);this.set(f,Kf(this,f,w,u))}connectedCallback(){if(S){if(this.debug=this.hasAttribute("debug"),this.debug)O(this,"Connected")}Object.entries(this.constructor.states).forEach(([f,u])=>{let w=c(u)?Kf(this,f,this.getAttribute(f)??void 0,void 0):u;this.set(f,w,!1)}),pu(this)}disconnectedCallback(){if(S&&this.debug)O(this,"Disconnected")}adoptedCallback(){if(S&&this.debug)O(this,"Adopted")}has(f){return this.signals.has(f)}get(f){let u=(J)=>!_f(J)?J:c(J)?u(J()):Gf(J)?u(J.get()):J,w=u(this.signals.get(f));if(S&&this.debug)O(w,`Get current value of state ${P(f)} in ${o(this)}`);return w}set(f,u,w=!0){let J;if(!this.signals.has(f)){if(S&&this.debug)J="Create";this.signals.set(f,If(u,!0))}else if(w){let h=this.signals.get(f);if(Gf(u)){if(S&&this.debug)J="Replace";if(this.signals.set(f,u),Cf(h))h.set(Hf)}else if(Cf(h)){if(S&&this.debug)J="Update";h.set(u)}else{O(u,`Computed state ${P(f)} in ${o(this)} cannot be set`,n);return}}else return;if(S&&this.debug)O(u,`${J} state ${P(f)} in ${o(this)}`)}delete(f){if(S&&this.debug)O(f,`Delete state ${P(f)} from ${o(this)}`);return this.signals.delete(f)}first(f){let u=this.root.querySelector(f);return new qf(this,u?[u]:[])}all(f){return new qf(this,Array.from(this.root.querySelectorAll(f)))}}var i=(f,u)=>(w,J)=>{let{read:h,update:X}=u,G=h(J);if(!c(f)){let $=Zf(f)&&Zf(G)?Kf(w,f,G):G;w.set(f,$,!1)}l(()=>{let $=h(J),z=c(f)?f($):w.get(f);if(!Object.is(z,$)){if(z===null&&u.delete)u.delete(J);else if(z==null&&G)X(J,G);else if(z!=null)X(J,z)}})};var lf=(f)=>i(f,{read:(u)=>u!=null,update:(u,w)=>w?xu(u):Promise.resolve(null)}),R=(f)=>i(f,{read:(u)=>u.textContent,update:(u,w)=>Lu(u,w)}),D=(f,u=f)=>i(u,{read:(w)=>w[f],update:(w,J)=>w[f]=J}),s=(f,u=f)=>i(u,{read:(w)=>w.getAttribute(f),update:(w,J)=>Iu(w,f,J),delete:(w)=>Nu(w,f)}),A=(f,u=f)=>i(u,{read:(w)=>w.hasAttribute(f),update:(w,J)=>Au(w,f,J)}),k=(f,u=f)=>i(u,{read:(w)=>w.classList.contains(f),update:(w,J)=>Tu(w,f,J)});class sf extends H{connectedCallback(){this.first("span").sync(R("name")),this.first("input").on("input",(f)=>{this.set("name",f.target?.value||void 0)})}}sf.define("hello-world");class kf extends H{static observedAttributes=["count"];static states={count:vf};connectedCallback(){this.set("parity",()=>(this.get("count")??0)%2?"odd":"even"),this.first(".increment").on("click",()=>{this.set("count",(f)=>f!=null?++f:1)}),this.first(".decrement").on("click",()=>{this.set("count",(f)=>f!=null?--f:0)}),this.first(".count").sync(R("count")),this.first(".parity").sync(R("parity"))}}kf.define("my-counter");var df=(f=0,u,w)=>(f+u+w)%w;class pf extends H{connectedCallback(){this.set("active",0);let f=this.querySelectorAll(".slide").length;this.first(".prev").on("click",()=>{this.set("active",(u)=>df(u,-1,f))}),this.first(".next").on("click",()=>{this.set("active",(u)=>df(u,1,f))}),this.all(".slide").sync((u,w,J)=>k("active",()=>J===this.get("active"))(u,w)),this.all(".dots span").sync((u,w,J)=>k("active",()=>J===this.get("active"))(u,w))}}pf.define("my-slider");class tf extends H{static observedAttributes=["disabled"];static states={disabled:_};connectedCallback(){this.first("button").sync(R("label"),D("disabled"))}}tf.define("input-button");class af extends H{static observedAttributes=["checked"];static states={checked:_};connectedCallback(){this.first("input").on("change",(f)=>{this.set("checked",Boolean(f.target?.checked))}).sync(D("checked")),this.self.sync(A("checked"))}}af.define("input-checkbox");class of extends H{static observedAttributes=["value"];connectedCallback(){this.self.sync(s("value")),this.all("input").on("change",(f)=>{this.set("value",f.target?.value)}),this.all("label").sync((f,u)=>k("selected",()=>f.get("value")===u.querySelector("input")?.value)(f,u))}}of.define("input-radiogroup");var Vf=(f)=>typeof f==="number",r=(f,u=!1,w=0)=>{let J=u?parseInt(f,10):parseFloat(f);return Number.isFinite(J)?J:w},mf=(f,u,w,J,h=!1)=>{let X=Math.round((w-u)/J),G=Math.round((f-u)*X/(w-u))/X;G=Math.min(Math.max(G,0),1);let $=G*(w-u)+u;return h?Math.round($):$};class nf extends H{static observedAttributes=["value","description"];static states={value:(f,u)=>u.isNumber?r(f,u.isInteger,0):f};isNumber;isInteger;input=null;step=1;min;max;connectedCallback(){if(this.input=this.querySelector("input"),this.isNumber=(this.input&&this.input.type==="number")??!1,this.isInteger=this.hasAttribute("integer"),this.input&&this.isNumber)this.step=r(this.input.step,this.isInteger,1),this.min=r(this.input.min,this.isInteger,0),this.max=r(this.input.max,this.isInteger,0);if(this.set("value",this.isNumber?this.input?.valueAsNumber:this.input?.value,!1),this.set("length",this.input?.value.length),this.set("empty",()=>!this.get("length")),this.first("input").on("change",()=>{this.#f(this.isNumber?this.input?.valueAsNumber:this.input?.value)}).on("input",()=>{this.set("length",this.input?.value.length)}),this.isNumber)this.first("input").on("keydown",(w)=>{let J=w;if(["ArrowUp","ArrowDown"].includes(J.key)){if(w.stopPropagation(),w.preventDefault(),J.key==="ArrowDown")this.stepDown(J.shiftKey?this.step*10:this.step);if(J.key==="ArrowUp")this.stepUp(J.shiftKey?this.step*10:this.step)}});this.set("ariaInvalid",()=>String(Boolean(this.get("error")))),this.set("aria-errormessage",()=>this.get("error")?this.querySelector(".error")?.id:void 0),this.first(".error").sync(R("error")),this.first("input").sync(D("ariaInvalid"),s("aria-errormessage"));let f=this.querySelector(".description");if(f){let w=this.input?.maxLength,J=w&&f.dataset.remaining,h=f.textContent;this.set("description",J?()=>{let X=this.get("length")??0;return X>0?J.replace("${x}",String(w-X)):h}:h),this.set("aria-describedby",()=>this.get("description")?f.id:void 0),this.first(".description").sync(R("description")),this.first("input").sync(s("aria-describedby"))}let u=this.querySelector(".spinbutton");if(!this.isNumber||!u)this.first(".decrement").on("click",(w)=>{this.stepDown(w.shiftKey?this.step*10:this.step)}).sync(D("disabled",()=>Vf(this.min)&&(this.get("value")??0-this.step<this.min))),this.first(".increment").on("click",(w)=>{this.stepUp(w.shiftKey?this.step*10:this.step)}).sync(D("disabled",()=>Vf(this.max)&&(this.get("value")??0+this.step>this.max)));this.first(".clear").on("click",()=>{this.clear(),this.input?.focus()}).sync(k("hidden","empty")),l(()=>{let w=this.get("value"),J=this.getAttribute("validate");if(w&&J)fetch(`${J}?name=${this.input?.name}value=${this.input?.value}`).then(async(h)=>{let X=await h.text();this.input?.setCustomValidity(X),this.set("error",X)}).catch((h)=>this.set("error",h.message));if(this.isNumber&&!Vf(w))return this.set("value",r(w,this.isInteger,0));if(this.input&&this.isNumber&&Number.isFinite(w))this.input.value=String(w)})}clear(){if(this.set("value",""),this.set("length",0),this.input)this.input.value="",this.input.focus()}stepUp(f=this.step){if(this.isNumber)this.#f((u)=>mf(u+f,this.min??0,this.max??100,this.step,this.isInteger))}stepDown(f=this.step){if(this.isNumber)this.#f((u)=>mf(u-f,this.min??0,this.max??100,this.step,this.isInteger))}#f=(f)=>{if(this.set("value",f),this.set("error",this.input?.validationMessage),typeof f==="function")f=this.get("value");if(this.input?.value!==String(f))this.dispatchEvent(new CustomEvent("value-change",{detail:f,bubbles:!0}))}}nf.define("input-field");var jf=Ku(rf(),1);(function(f){var u="\\b(?:BASH|BASHOPTS|BASH_ALIASES|BASH_ARGC|BASH_ARGV|BASH_CMDS|BASH_COMPLETION_COMPAT_DIR|BASH_LINENO|BASH_REMATCH|BASH_SOURCE|BASH_VERSINFO|BASH_VERSION|COLORTERM|COLUMNS|COMP_WORDBREAKS|DBUS_SESSION_BUS_ADDRESS|DEFAULTS_PATH|DESKTOP_SESSION|DIRSTACK|DISPLAY|EUID|GDMSESSION|GDM_LANG|GNOME_KEYRING_CONTROL|GNOME_KEYRING_PID|GPG_AGENT_INFO|GROUPS|HISTCONTROL|HISTFILE|HISTFILESIZE|HISTSIZE|HOME|HOSTNAME|HOSTTYPE|IFS|INSTANCE|JOB|LANG|LANGUAGE|LC_ADDRESS|LC_ALL|LC_IDENTIFICATION|LC_MEASUREMENT|LC_MONETARY|LC_NAME|LC_NUMERIC|LC_PAPER|LC_TELEPHONE|LC_TIME|LESSCLOSE|LESSOPEN|LINES|LOGNAME|LS_COLORS|MACHTYPE|MAILCHECK|MANDATORY_PATH|NO_AT_BRIDGE|OLDPWD|OPTERR|OPTIND|ORBIT_SOCKETDIR|OSTYPE|PAPERSIZE|PATH|PIPESTATUS|PPID|PS1|PS2|PS3|PS4|PWD|RANDOM|REPLY|SECONDS|SELINUX_INIT|SESSION|SESSIONTYPE|SESSION_MANAGER|SHELL|SHELLOPTS|SHLVL|SSH_AUTH_SOCK|TERM|UID|UPSTART_EVENTS|UPSTART_INSTANCE|UPSTART_JOB|UPSTART_SESSION|USER|WINDOWID|XAUTHORITY|XDG_CONFIG_DIRS|XDG_CURRENT_DESKTOP|XDG_DATA_DIRS|XDG_GREETER_DATA_DIR|XDG_MENU_PREFIX|XDG_RUNTIME_DIR|XDG_SEAT|XDG_SEAT_PATH|XDG_SESSION_DESKTOP|XDG_SESSION_ID|XDG_SESSION_PATH|XDG_SESSION_TYPE|XDG_VTNR|XMODIFIERS)\\b",w={pattern:/(^(["']?)\w+\2)[ \t]+\S.*/,lookbehind:!0,alias:"punctuation",inside:null},J={bash:w,environment:{pattern:RegExp("\\$"+u),alias:"constant"},variable:[{pattern:/\$?\(\([\s\S]+?\)\)/,greedy:!0,inside:{variable:[{pattern:/(^\$\(\([\s\S]+)\)\)/,lookbehind:!0},/^\$\(\(/],number:/\b0x[\dA-Fa-f]+\b|(?:\b\d+(?:\.\d*)?|\B\.\d+)(?:[Ee]-?\d+)?/,operator:/--|\+\+|\*\*=?|<<=?|>>=?|&&|\|\||[=!+\-*/%<>^&|]=?|[?~:]/,punctuation:/\(\(?|\)\)?|,|;/}},{pattern:/\$\((?:\([^)]+\)|[^()])+\)|`[^`]+`/,greedy:!0,inside:{variable:/^\$\(|^`|\)$|`$/}},{pattern:/\$\{[^}]+\}/,greedy:!0,inside:{operator:/:[-=?+]?|[!\/]|##?|%%?|\^\^?|,,?/,punctuation:/[\[\]]/,environment:{pattern:RegExp("(\\{)"+u),lookbehind:!0,alias:"constant"}}},/\$(?:\w+|[#?*!@$])/],entity:/\\(?:[abceEfnrtv\\"]|O?[0-7]{1,3}|U[0-9a-fA-F]{8}|u[0-9a-fA-F]{4}|x[0-9a-fA-F]{1,2})/};f.languages.bash={shebang:{pattern:/^#!\s*\/.*/,alias:"important"},comment:{pattern:/(^|[^"{\\$])#.*/,lookbehind:!0},"function-name":[{pattern:/(\bfunction\s+)[\w-]+(?=(?:\s*\(?:\s*\))?\s*\{)/,lookbehind:!0,alias:"function"},{pattern:/\b[\w-]+(?=\s*\(\s*\)\s*\{)/,alias:"function"}],"for-or-select":{pattern:/(\b(?:for|select)\s+)\w+(?=\s+in\s)/,alias:"variable",lookbehind:!0},"assign-left":{pattern:/(^|[\s;|&]|[<>]\()\w+(?:\.\w+)*(?=\+?=)/,inside:{environment:{pattern:RegExp("(^|[\\s;|&]|[<>]\\()"+u),lookbehind:!0,alias:"constant"}},alias:"variable",lookbehind:!0},parameter:{pattern:/(^|\s)-{1,2}(?:\w+:[+-]?)?\w+(?:\.\w+)*(?=[=\s]|$)/,alias:"variable",lookbehind:!0},string:[{pattern:/((?:^|[^<])<<-?\s*)(\w+)\s[\s\S]*?(?:\r?\n|\r)\2/,lookbehind:!0,greedy:!0,inside:J},{pattern:/((?:^|[^<])<<-?\s*)(["'])(\w+)\2\s[\s\S]*?(?:\r?\n|\r)\3/,lookbehind:!0,greedy:!0,inside:{bash:w}},{pattern:/(^|[^\\](?:\\\\)*)"(?:\\[\s\S]|\$\([^)]+\)|\$(?!\()|`[^`]+`|[^"\\`$])*"/,lookbehind:!0,greedy:!0,inside:J},{pattern:/(^|[^$\\])'[^']*'/,lookbehind:!0,greedy:!0},{pattern:/\$'(?:[^'\\]|\\[\s\S])*'/,greedy:!0,inside:{entity:J.entity}}],environment:{pattern:RegExp("\\$?"+u),alias:"constant"},variable:J.variable,function:{pattern:/(^|[\s;|&]|[<>]\()(?:add|apropos|apt|apt-cache|apt-get|aptitude|aspell|automysqlbackup|awk|basename|bash|bc|bconsole|bg|bzip2|cal|cargo|cat|cfdisk|chgrp|chkconfig|chmod|chown|chroot|cksum|clear|cmp|column|comm|composer|cp|cron|crontab|csplit|curl|cut|date|dc|dd|ddrescue|debootstrap|df|diff|diff3|dig|dir|dircolors|dirname|dirs|dmesg|docker|docker-compose|du|egrep|eject|env|ethtool|expand|expect|expr|fdformat|fdisk|fg|fgrep|file|find|fmt|fold|format|free|fsck|ftp|fuser|gawk|git|gparted|grep|groupadd|groupdel|groupmod|groups|grub-mkconfig|gzip|halt|head|hg|history|host|hostname|htop|iconv|id|ifconfig|ifdown|ifup|import|install|ip|java|jobs|join|kill|killall|less|link|ln|locate|logname|logrotate|look|lpc|lpr|lprint|lprintd|lprintq|lprm|ls|lsof|lynx|make|man|mc|mdadm|mkconfig|mkdir|mke2fs|mkfifo|mkfs|mkisofs|mknod|mkswap|mmv|more|most|mount|mtools|mtr|mutt|mv|nano|nc|netstat|nice|nl|node|nohup|notify-send|npm|nslookup|op|open|parted|passwd|paste|pathchk|ping|pkill|pnpm|podman|podman-compose|popd|pr|printcap|printenv|ps|pushd|pv|quota|quotacheck|quotactl|ram|rar|rcp|reboot|remsync|rename|renice|rev|rm|rmdir|rpm|rsync|scp|screen|sdiff|sed|sendmail|seq|service|sftp|sh|shellcheck|shuf|shutdown|sleep|slocate|sort|split|ssh|stat|strace|su|sudo|sum|suspend|swapon|sync|sysctl|tac|tail|tar|tee|time|timeout|top|touch|tr|traceroute|tsort|tty|umount|uname|unexpand|uniq|units|unrar|unshar|unzip|update-grub|uptime|useradd|userdel|usermod|users|uudecode|uuencode|v|vcpkg|vdir|vi|vim|virsh|vmstat|wait|watch|wc|wget|whereis|which|who|whoami|write|xargs|xdg-open|yarn|yes|zenity|zip|zsh|zypper)(?=$|[)\s;|&])/,lookbehind:!0},keyword:{pattern:/(^|[\s;|&]|[<>]\()(?:case|do|done|elif|else|esac|fi|for|function|if|in|select|then|until|while)(?=$|[)\s;|&])/,lookbehind:!0},builtin:{pattern:/(^|[\s;|&]|[<>]\()(?:\.|:|alias|bind|break|builtin|caller|cd|command|continue|declare|echo|enable|eval|exec|exit|export|getopts|hash|help|let|local|logout|mapfile|printf|pwd|read|readarray|readonly|return|set|shift|shopt|source|test|times|trap|type|typeset|ulimit|umask|unalias|unset)(?=$|[)\s;|&])/,lookbehind:!0,alias:"class-name"},boolean:{pattern:/(^|[\s;|&]|[<>]\()(?:false|true)(?=$|[)\s;|&])/,lookbehind:!0},"file-descriptor":{pattern:/\B&\d\b/,alias:"important"},operator:{pattern:/\d?<>|>\||\+=|=[=~]?|!=?|<<[<-]?|[&\d]?>>|\d[<>]&?|[<>][&=]?|&[>&]?|\|[&|]?/,inside:{"file-descriptor":{pattern:/^\d/,alias:"important"}}},punctuation:/\$?\(\(?|\)\)?|\.\.|[{}[\];\\]/,number:{pattern:/(^|\s)(?:[1-9]\d*|0)(?:[.,]\d+)?\b/,lookbehind:!0}},w.inside=f.languages.bash;var h=["comment","function-name","for-or-select","assign-left","parameter","string","environment","function","keyword","builtin","boolean","file-descriptor","operator","punctuation","number"],X=J.variable[1].inside;for(var G=0;G<h.length;G++)X[h[G]]=f.languages.bash[h[G]];f.languages.sh=f.languages.bash,f.languages.shell=f.languages.bash})(Prism);Prism.languages.json={property:{pattern:/(^|[^\\])"(?:\\.|[^\\"\r\n])*"(?=\s*:)/,lookbehind:!0,greedy:!0},string:{pattern:/(^|[^\\])"(?:\\.|[^\\"\r\n])*"(?!\s*:)/,lookbehind:!0,greedy:!0},comment:{pattern:/\/\/.*|\/\*[\s\S]*?(?:\*\/|$)/,greedy:!0},number:/-?\b\d+(?:\.\d+)?(?:e[+-]?\d+)?\b/i,punctuation:/[{}[\],]/,operator:/:/,boolean:/\b(?:false|true)\b/,null:{pattern:/\bnull\b/,alias:"keyword"}};Prism.languages.webmanifest=Prism.languages.json;(function(f){f.languages.typescript=f.languages.extend("javascript",{"class-name":{pattern:/(\b(?:class|extends|implements|instanceof|interface|new|type)\s+)(?!keyof\b)(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*(?:\s*<(?:[^<>]|<(?:[^<>]|<[^<>]*>)*>)*>)?/,lookbehind:!0,greedy:!0,inside:null},builtin:/\b(?:Array|Function|Promise|any|boolean|console|never|number|string|symbol|unknown)\b/}),f.languages.typescript.keyword.push(/\b(?:abstract|declare|is|keyof|readonly|require)\b/,/\b(?:asserts|infer|interface|module|namespace|type)\b(?=\s*(?:[{_$a-zA-Z\xA0-\uFFFF]|$))/,/\btype\b(?=\s*(?:[\{*]|$))/),delete f.languages.typescript.parameter,delete f.languages.typescript["literal-property"];var u=f.languages.extend("typescript",{});delete u["class-name"],f.languages.typescript["class-name"].inside=u,f.languages.insertBefore("typescript","function",{decorator:{pattern:/@[$\w\xA0-\uFFFF]+/,inside:{at:{pattern:/^@/,alias:"operator"},function:/^[\s\S]+/}},"generic-function":{pattern:/#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*\s*<(?:[^<>]|<(?:[^<>]|<[^<>]*>)*>)*>(?=\s*\()/,greedy:!0,inside:{function:/^#?(?!\s)[_$a-zA-Z\xA0-\uFFFF](?:(?!\s)[$\w\xA0-\uFFFF])*/,generic:{pattern:/<[\s\S]+/,alias:"class-name",inside:u}}}}),f.languages.ts=f.languages.typescript})(Prism);class ef extends H{static observedAttributes=["collapsed"];static states={collapsed:_};connectedCallback(){let f=this.getAttribute("language")||"html",u=this.querySelector("code");if(u)this.set("code",u.textContent?.trim(),!1),l(()=>{let w=document.createElement("code");w.innerHTML=jf.default.highlight(this.get("code")??"",jf.default.languages[f],f),N(()=>{Array.from(w.childNodes).filter((J)=>J.nodeType!==Node.COMMENT_NODE).forEach((J)=>J.remove()),Array.from(w.childNodes).forEach((J)=>w.appendChild(J))},[w,"h"])}),this.first(".copy").on("click",async(w)=>{let J=w.currentTarget,h=J.textContent,X="success";try{await navigator.clipboard.writeText(u.textContent??"")}catch(G){console.error("Error when trying to use navigator.clipboard.writeText()",G),X="error"}J.set("disabled",!0),J.set("label",this.getAttribute(`copy-${X}`)),setTimeout(()=>{J.set("disabled",!1),J.set("label",h)},X==="success"?1000:3000)}),this.first(".overlay").on("click",()=>this.set("collapsed",!1)),this.self.sync(A("collapsed"))}}ef.define("code-block");class fu extends H{connectedCallback(){this.set("open",this.hasAttribute("open"),!1),this.set("collapsible",this.hasAttribute("collapsible"),!1),this.self.sync(A("open"),A("collapsible"),D("ariaHidden",()=>!this.get("open")&&!this.get("collapsible"))),this.first("details").sync(D("open"),D("ariaDisabled",()=>!this.get("collapsible")))}}fu.define("accordion-panel");class uu extends H{static observedAttributes=["accordion"];static states={accordion:_};static consumedContexts=["media-viewport"];connectedCallback(){super.connectedCallback(),this.set("active",0,!1),setTimeout(()=>{if(this.get("media-viewport"))this.set("accordion",()=>["xs","sm"].includes(String(this.get("media-viewport"))))},0),this.self.sync(A("accordion")),this.first("menu").sync(D("ariaHidden","accordion")),this.all("menu button").on("click",(f,u)=>()=>this.set("active",u)).sync((f,u,w)=>D("ariaPressed",()=>this.get("active")===w)(f,u)),this.all("accordion-panel").pass({open:(f,u)=>()=>this.get("active")===u,collapsible:"accordion"})}}uu.define("tab-list");class wu extends H{static states={content:"",error:"",loaded:!1};connectedCallback(){super.connectedCallback();let f=this.getAttribute("src");if(f)this.loadContent(f).then(([u,w])=>{this.set("content",w),this.set("error",u),this.set("loaded",!0)});this.first(".loading").sync(lf("loaded")),this.first(".error").sync(R("error")),l(()=>{let u=this.get("content");if(u)N(()=>{this.root.innerHTML=u,this.root.querySelectorAll("script").forEach((w)=>{let J=document.createElement("script");J.appendChild(document.createTextNode(w.textContent??"")),this.root.appendChild(J),w.remove()})},[this.root,"h"])})}async loadContent(f){try{let u=await fetch(f);if(!u.ok)return[u.statusText,""];return["",await u.text()]}catch(u){return[u.message,""]}}}wu.define("lazy-load");var Yf=(f,u)=>{let w=f?.trim();if(!w)return u;let J=w.match(/em$/)?"em":"px",h=parseFloat(w);return Number.isFinite(h)?h+J:u},Df="media-motion",Ff="media-theme",t="media-viewport",Ju="media-orientation";class hu extends H{static providedContexts=[Df,Ff,t,Ju];connectedCallback(){super.connectedCallback();let f="light",u="dark",w="xs",J="sm",h="md",X="lg",G="xl",$="landscape",z="portrait",j=(()=>({sm:Yf(this.getAttribute(J),"32em"),md:Yf(this.getAttribute(h),"48em"),lg:Yf(this.getAttribute(X),"72em"),xl:Yf(this.getAttribute(G),"108em")}))(),v=matchMedia("(prefers-reduced-motion: reduce)"),B=matchMedia("(prefers-color-scheme: dark)"),K=matchMedia(`(min-width: ${j.sm})`),V=matchMedia(`(min-width: ${j.md})`),Z=matchMedia(`(min-width: ${j.lg})`),q=matchMedia(`(min-width: ${j.xl})`),Q=matchMedia("(orientation: landscape)"),Y=()=>{if(q.matches)return G;if(Z.matches)return X;if(V.matches)return h;if(K.matches)return J;return w};this.set(Df,v.matches),this.set(Ff,B.matches?u:f),this.set(t,Y()),this.set(Ju,Q.matches?$:z),v.addEventListener("change",(F)=>this.set(Df,F.matches)),B.addEventListener("change",(F)=>this.set(Ff,F.matches?u:f)),K.addEventListener("change",()=>this.set(t,Y())),V.addEventListener("change",()=>this.set(t,Y())),Z.addEventListener("change",()=>this.set(t,Y())),q.addEventListener("change",()=>this.set(t,Y())),Q.addEventListener("change",(F)=>{this.set(Ff,F.matches?$:z)})}}hu.define("media-context");class qu extends H{connectedCallback(){let f=this.querySelector("input-field");this.first("form").on("submit",(u)=>{u.preventDefault(),setTimeout(()=>{this.dispatchEvent(new CustomEvent("add-todo",{bubbles:!0,detail:f?.get("value")??""})),f?.clear()},0)}),this.first("input-button").pass({disabled:()=>f?.get("empty")})}}qu.define("todo-form");class Zu extends H{static states={tasks:[],filter:"all"};connectedCallback(){this.#f(),this.self.on("click",(f)=>{if(f.target.localName==="button")this.removeItem(f.target)}).sync(s("filter")),this.set("count",()=>{let f=this.get("tasks").map((J)=>J.signals.get("checked")),u=f.filter((J)=>J?.get()).length,w=f.length;return{active:w-u,completed:u,total:w}})}addItem=(f)=>{let u=this.querySelector("template")?.content.cloneNode(!0);if(u&&u.querySelector("span"))u.querySelector("span").textContent=f,this.querySelector("ol")?.appendChild(u),this.#f()};removeItem=(f)=>{f.closest("li")?.remove(),this.#f()};clearCompleted=()=>{this.get("tasks").filter((f)=>f.get("checked")).forEach((f)=>f.parentElement?.remove()),this.#f()};#f(){this.set("tasks",Array.from(this.querySelectorAll("input-checkbox")))}}Zu.define("todo-list");class Qu extends H{static states={active:0};connectedCallback(){this.first(".count").sync(R("active")),this.first(".singular").sync(D("ariaHidden",()=>this.get("active")>1)),this.first(".plural").sync(D("ariaHidden",()=>this.get("active")===1)),this.first(".remaining").sync(D("ariaHidden",()=>!this.get("active"))),this.first(".all-done").sync(D("ariaHidden",()=>!!this.get("active")))}}Qu.define("todo-count");class Yu extends H{connectedCallback(){let f=this.querySelector("todo-list"),u=this.querySelector("input-radiogroup");this.self.on("add-todo",(w)=>{f?.addItem(w.detail)}),this.first("todo-count").pass({active:()=>(f?.get("count")).active}),this.first("todo-list").pass({filter:()=>u?.get("value")}),this.first(".clear-completed").on("click",()=>f?.clearCompleted()).pass({disabled:()=>!(f?.get("count")).completed})}}Yu.define("todo-app");
