+++ index.js
@@ -1,1 +1,1 @@
-var T=($)=>typeof $==="function",n=($)=>T($)&&/^async\s+/.test($.toString()),R=($)=>(A)=>A instanceof $,d=R(Error),u=R(Promise);var l="Computed",G=($,A)=>{A=A??n($);let B=[],K,Z=null,W=!0,J=()=>{if(W=!0,A)b(B)},P={[Symbol.toStringTag]:l,get:()=>{if(A)O(B);if(!A||W)k(()=>{let j=(H)=>{K=H,W=!1,Z=null},v=(H)=>{Z=d(H)?H:new Error(`Computed function failed: ${H}`)};try{let H=$(K);u(H)?H.then(j).catch(v):j(H)}catch(H){v(H)}},J);if(d(Z))throw Z;return K},map:(j)=>G(()=>j(P.get()))};return P},i=($)=>!!$&&typeof $==="object"&&$[Symbol.toStringTag]===l;var F,o=!1,c=[],I=($)=>L($)||i($),U=($,A=!1)=>I($)?$:T($)?G($,A):_($),O=($)=>{if(F&&!$.includes(F))$.push(F)},b=($)=>$.forEach((A)=>o?c.push(A):A()),k=($,A)=>{let B=F;F=A,$(),F=B},t=($)=>{o=!0,$(),o=!1,c.forEach((A)=>A()),c.length=0};var V=Symbol();class E{$;watchers=[];constructor($){this.value=$}get(){return O(this.watchers),this.value}set($){if(V!==$){let A=T($)?$(this.value):$;if(Object.is(this.value,A))return;this.value=A}if(b(this.watchers),V===$)this.watchers=[]}map($){return G(()=>$(this.get()))}}var _=($)=>new E($),L=R(E);var D=($)=>{let A=()=>k(()=>{try{$()}catch(B){console.error(B)}},A);A()};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=($)=>setTimeout($,16);var h=new Map,f=[],g,a=()=>{g=null,f.forEach(($)=>$()),f=[],h.clear()},N$=()=>{if(g)cancelAnimationFrame(g);g=requestAnimationFrame(a)};queueMicrotask(a);var X=($,A)=>new Promise((B,K)=>{let Z=()=>{try{B($())}catch(W){K(W)}};if(A){let[W,J]=A;if(!h.has(W))h.set(W,new Map);let P=h.get(W);if(P.has(J)){let j=f.indexOf($);if(j>-1)f.splice(j,1)}P.set(J,Z)}f.push(Z),N$()}),e=async()=>new Promise(requestAnimationFrame);var $$=($)=>$.nodeType===Node.COMMENT_NODE,G$=($)=>!/^on/i.test($),L$=($)=>{if(/^(mailto|tel):/i.test($))return!0;if($.includes("://"))try{let A=new URL($,window.location.origin);return!["http:","https:","ftp:"].includes(A.protocol)}catch(A){return!0}return!0},m=($,A,B)=>{if(!G$(A))throw new Error(`Unsafe attribute: ${A}`);if(B=String(B).trim(),!L$(B))throw new Error(`Unsafe URL for ${A}: ${B}`);$.setAttribute(A,B)};var A$=($,A,B={},K)=>X(()=>{let Z=document.createElement(A);for(let[W,J]of Object.entries(B))m(Z,W,J);if(K)Z.textContent=K;return $.append(Z),Z},[$,"e"]),B$=($)=>X(()=>{return $.remove(),null},[$,"r"]),K$=($,A)=>X(()=>{return Array.from($.childNodes).filter((B)=>!$$(B)).forEach((B)=>B.remove()),$.append(document.createTextNode(A)),$},[$,"t"]),Z$=($,A,B)=>X(()=>{return m($,A,B),$},[$,`a:${A}`]),W$=($,A)=>X(()=>{return $.removeAttribute(A),$},[$,`a:${A}`]),J$=($,A,B)=>X(()=>{return $.toggleAttribute(A,B),$},[$,`a:${A}`]),Y$=($,A,B)=>X(()=>{return $.classList.toggle(A,B),$},[$,`c:${A}`]),j$=($,A,B)=>X(()=>{return $.style.setProperty(A,B),$},[$,`s:${A}`]),Q$=($,A)=>X(()=>{return $.style.removeProperty(A),$},[$,`s:${A}`]);var z=($)=>typeof $==="function",p=($)=>!!$&&typeof $==="object",F$=($)=>typeof $==="number",y=($)=>typeof $==="string",I$=($)=>typeof $==="symbol",X$=($)=>y($)||I$($)||F$($);var Q=!1,z$="debug",U$="info",H$="warn",w="error",V$=($)=>$?`#${$}`:"",y$=($)=>$.length?`.${Array.from($).join(".")}`:"",x=($)=>`<${$.localName}${V$($.id)}${y$($.classList)}>`,C=($)=>y($)?`"${$}"`:p($)?JSON.stringify($):String($),Y=($,A,B=z$)=>{if(Q||[w,H$].includes(B))console[B](A,$);return $};var x$=($)=>z($)&&$.length===2,s=($,A,B=0)=>x$($)?$(A,B):$;class q{$;A;constructor($,A=[$]){this.host=$;this.targets=A}on($,A){return this.targets.forEach((B,K)=>B.addEventListener($,s(A,B,K))),this}off($,A){return this.targets.forEach((B,K)=>B.removeEventListener($,s(A,B,K))),this}pass($){return this.targets.forEach(async(A,B)=>{if(await S.registry.whenDefined(A.localName),A instanceof S)Object.entries($).forEach(([K,Z])=>{let W=s(Z,A,B),J=X$(W)?this.host.signals.get(W):U(W,!0);if(J)A.set(K,J);else Y(Z,`Invalid source for state ${C(K)}`,w)});else Y(A,"Target is not a UIElement",w)}),this}sync(...$){return this.targets.forEach((A,B)=>$.forEach((K)=>K(this.host,A,B))),this}}var P$=($,A)=>{if(A==null)return;let B=$(A);return Number.isFinite(B)?B:void 0},M=($,A,B,K=void 0)=>{let Z=$.constructor.states[A];return z(Z)&&!!Z.length?Z(B,$,K):B},S$=($)=>$!=null,q$=($)=>P$(parseInt,$),M$=($)=>P$(parseFloat,$),T$=($)=>$,D$=($)=>(A)=>{if(A==null)return;return $.includes(A.toLowerCase())?A:void 0},f$=($)=>{if($==null)return;try{return JSON.parse($)}catch(A){Y(A,"Failed to parse JSON",w);return}};var w$="context-request";class C$ extends Event{$;A;B;constructor($,A,B=!1){super(w$,{bubbles:!0,composed:!0});this.context=$;this.callback=A;this.subscribe=B}}var r=($)=>{let A=$.constructor,B=A.consumedContexts||[];setTimeout(()=>{for(let Z of B)$.dispatchEvent(new C$(Z,(W)=>$.set(String(Z),W)))});let K=A.providedContexts||[];if(!K.length)return!1;return $.addEventListener(w$,(Z)=>{let{context:W,callback:J}=Z;if(!K.includes(W)||!z(J))return;Z.stopPropagation(),J($.signals.get(String(W)))}),!0};class S extends HTMLElement{static registry=customElements;static states={};static observedAttributes;static consumedContexts;static providedContexts;static define($){try{if(S.registry.define($,this),Q)Y($,"Registered custom element")}catch(A){Y(A,`Failed to register custom element ${$}`,w)}}signals=new Map;self=new q(this);root=this.shadowRoot||this;debug=!1;attributeChangedCallback($,A,B){if(B===A)return;if(Q&&this.debug)Y(`${C(A)} => ${C(B)}`,`Attribute "${$}" of ${x(this)} changed`);this.set($,M(this,$,B,A))}connectedCallback(){if(Q){if(this.debug=this.hasAttribute("debug"),this.debug)Y(this,"Connected")}Object.entries(this.constructor.states).forEach(([$,A])=>{let B=z(A)?M(this,$,this.getAttribute($)??void 0,void 0):A;this.set($,B,!1)}),r(this)}disconnectedCallback(){if(Q&&this.debug)Y(this,"Disconnected")}adoptedCallback(){if(Q&&this.debug)Y(this,"Adopted")}has($){return this.signals.has($)}get($){let A=(K)=>!p(K)?K:z(K)?A(K()):I(K)?A(K.get()):K,B=A(this.signals.get($));if(Q&&this.debug)Y(B,`Get current value of state ${C($)} in ${x(this)}`);return B}set($,A,B=!0){let K;if(!this.signals.has($)){if(Q&&this.debug)K="Create";this.signals.set($,U(A,!0))}else if(B){let Z=this.signals.get($);if(I(A)){if(Q&&this.debug)K="Replace";if(this.signals.set($,A),L(Z))Z.set(V)}else if(L(Z)){if(Q&&this.debug)K="Update";Z.set(A)}else{Y(A,`Computed state ${C($)} in ${x(this)} cannot be set`,w);return}}else return;if(Q&&this.debug)Y(A,`${K} state ${C($)} in ${x(this)}`)}delete($){if(Q&&this.debug)Y($,`Delete state ${C($)} from ${x(this)}`);return this.signals.delete($)}first($){let A=this.root.querySelector($);return new q(this,A?[A]:[])}all($){return new q(this,Array.from(this.root.querySelectorAll($)))}}var R$=($,A=$)=>(B,K)=>D(()=>{K.dispatchEvent(new CustomEvent($,{detail:B.get(A),bubbles:!0}))}),N=($,A)=>(B,K)=>{let{read:Z,update:W}=A,J=Z(K);if(!z($)){let P=y($)&&y(J)?M(B,$,J):J;B.set($,P,!1)}D(()=>{let P=Z(K),j=z($)?$(P):B.get($);if(!Object.is(j,P)){if(j===null&&A.delete)A.delete(K);else if(j==null&&J)W(K,J);else if(j!=null)W(K,j)}})},O$=($,A)=>N(A,{read:()=>null,update:(B,K)=>A$(B,$,K)}),b$=($)=>N($,{read:(A)=>A!=null,update:(A,B)=>B?B$(A):Promise.resolve(null)}),k$=($)=>N($,{read:(A)=>A.textContent,update:(A,B)=>K$(A,B)}),_$=($,A=$)=>N(A,{read:(B)=>B[$],update:(B,K)=>B[$]=K}),E$=($,A=$)=>N(A,{read:(B)=>B.getAttribute($),update:(B,K)=>Z$(B,$,K),delete:(B)=>W$(B,$)}),h$=($,A=$)=>N(A,{read:(B)=>B.hasAttribute($),update:(B,K)=>J$(B,$,K)}),g$=($,A=$)=>N(A,{read:(B)=>B.classList.contains($),update:(B,K)=>Y$(B,$,K)}),p$=($,A=$)=>N(A,{read:(B)=>B.style.getPropertyValue($),update:(B,K)=>j$(B,$,K),delete:(B)=>Q$(B,$)});export{r as useContext,N as updateElement,g$ as toggleClass,h$ as toggleAttribute,U as toSignal,_ as state,k$ as setText,p$ as setStyle,_$ as setProperty,E$ as setAttribute,b$ as removeElement,M as parse,Y as log,L as isState,I as isSignal,X as enqueue,R$ as emit,D as effect,O$ as createElement,G as computed,t as batch,T$ as asString,M$ as asNumber,f$ as asJSON,q$ as asInteger,D$ as asEnum,S$ as asBoolean,e as animationFrame,V as UNSET,S as UIElement,q as UI,H$ as LOG_WARN,U$ as LOG_INFO,w as LOG_ERROR,z$ as LOG_DEBUG};
+var T=($)=>typeof $==="function",n=($)=>T($)&&/^async\s+/.test($.toString()),R=($)=>(A)=>A instanceof $,p=R(Error),u=R(Promise);var l="Computed",G=($,A)=>{A=A??n($);let B=[],K,Z=null,W=!0,J=()=>{if(W=!0,A)b(B)},P={[Symbol.toStringTag]:l,get:()=>{if(A)O(B);if(!A||W)k(()=>{let j=(H)=>{K=H,W=!1,Z=null},v=(H)=>{Z=p(H)?H:new Error(`Computed function failed: ${H}`)};try{let H=$(K);u(H)?H.then(j).catch(v):j(H)}catch(H){v(H)}},J);if(p(Z))throw Z;return K},map:(j)=>G(()=>j(P.get()))};return P},i=($)=>!!$&&typeof $==="object"&&$[Symbol.toStringTag]===l;var F,o=!1,c=[],I=($)=>L($)||i($),U=($,A=!1)=>I($)?$:T($)?G($,A):_($),O=($)=>{if(F&&!$.includes(F))$.push(F)},b=($)=>$.forEach((A)=>o?c.push(A):A()),k=($,A)=>{let B=F;F=A,$(),F=B},t=($)=>{o=!0,$(),o=!1,c.forEach((A)=>A()),c.length=0};var V=Symbol();class E{$;watchers=[];constructor($){this.value=$}get(){return O(this.watchers),this.value}set($){if(V!==$){let A=T($)?$(this.value):$;if(Object.is(this.value,A))return;this.value=A}if(b(this.watchers),V===$)this.watchers=[]}map($){return G(()=>$(this.get()))}}var _=($)=>new E($),L=R(E);var D=($)=>{let A=()=>k(()=>{try{$()}catch(B){console.error(B)}},A);A()};if(!("requestAnimationFrame"in globalThis))globalThis.requestAnimationFrame=($)=>setTimeout($,16);var h=new Map,f=[],g,a=()=>{g=null,f.forEach(($)=>$()),f=[],h.clear()},N$=()=>{if(g)cancelAnimationFrame(g);g=requestAnimationFrame(a)};queueMicrotask(a);var X=($,A)=>new Promise((B,K)=>{let Z=()=>{try{B($())}catch(W){K(W)}};if(A){let[W,J]=A;if(!h.has(W))h.set(W,new Map);let P=h.get(W);if(P.has(J)){let j=f.indexOf($);if(j>-1)f.splice(j,1)}P.set(J,Z)}f.push(Z),N$()}),e=async()=>new Promise(requestAnimationFrame);var $$=($)=>$.nodeType===Node.COMMENT_NODE,G$=($)=>!/^on/i.test($),L$=($)=>{if(/^(mailto|tel):/i.test($))return!0;if($.includes("://"))try{let A=new URL($,window.location.origin);return!["http:","https:","ftp:"].includes(A.protocol)}catch(A){return!0}return!0},m=($,A,B)=>{if(!G$(A))throw new Error(`Unsafe attribute: ${A}`);if(B=String(B).trim(),!L$(B))throw new Error(`Unsafe URL for ${A}: ${B}`);$.setAttribute(A,B)};var A$=($,A,B={},K)=>X(()=>{let Z=document.createElement(A);for(let[W,J]of Object.entries(B))m(Z,W,J);if(K)Z.textContent=K;return $.append(Z),Z},[$,"e"]),B$=($)=>X(()=>{return $.remove(),null},[$,"r"]),K$=($,A)=>X(()=>{return Array.from($.childNodes).filter((B)=>!$$(B)).forEach((B)=>B.remove()),$.append(document.createTextNode(A)),$},[$,"t"]),Z$=($,A,B)=>X(()=>{return m($,A,B),$},[$,`a:${A}`]),W$=($,A)=>X(()=>{return $.removeAttribute(A),$},[$,`a:${A}`]),J$=($,A,B)=>X(()=>{return $.toggleAttribute(A,B),$},[$,`a:${A}`]),Y$=($,A,B)=>X(()=>{return $.classList.toggle(A,B),$},[$,`c:${A}`]),j$=($,A,B)=>X(()=>{return $.style.setProperty(A,B),$},[$,`s:${A}`]),Q$=($,A)=>X(()=>{return $.style.removeProperty(A),$},[$,`s:${A}`]);var z=($)=>typeof $==="function",d=($)=>!!$&&typeof $==="object",F$=($)=>typeof $==="number",x=($)=>typeof $==="string",I$=($)=>typeof $==="symbol",X$=($)=>x($)||I$($)||F$($);var Q=!1,z$="debug",U$="info",H$="warn",C="error",V$=($)=>$?`#${$}`:"",x$=($)=>$.length?`.${Array.from($).join(".")}`:"",S=($)=>`<${$.localName}${V$($.id)}${x$($.classList)}>`,w=($)=>x($)?`"${$}"`:d($)?JSON.stringify($):String($),Y=($,A,B=z$)=>{if(Q||[C,H$].includes(B))console[B](A,$);return $};var S$=($)=>z($)&&$.length===2,s=($,A,B=0)=>S$($)?$(A,B):$;class y{$;A;constructor($,A=[$]){this.host=$;this.targets=A}on($,A){return this.targets.forEach((B,K)=>B.addEventListener($,s(A,B,K))),this}off($,A){return this.targets.forEach((B,K)=>B.removeEventListener($,s(A,B,K))),this}pass($){return this.targets.forEach(async(A,B)=>{if(await q.registry.whenDefined(A.localName),A instanceof q)Object.entries($).forEach(([K,Z])=>{let W=s(Z,A,B),J=X$(W)?this.host.signals.get(W):U(W,!0);if(J)A.set(K,J);else Y(Z,`Invalid source for state ${w(K)}`,C)});else Y(A,"Target is not a UIElement",C)}),this}sync(...$){return this.targets.forEach((A,B)=>$.forEach((K)=>K(this.host,A,B))),this}}var P$=($,A)=>{if(A==null)return;let B=$(A);return Number.isFinite(B)?B:void 0},M=($,A,B,K=void 0)=>{let Z=$.constructor.states[A];return z(Z)&&!!Z.length?Z(B,$,K):B},q$=($)=>$!=null,y$=($)=>P$(parseInt,$),M$=($)=>P$(parseFloat,$),T$=($)=>$,D$=($)=>(A)=>{if(A==null)return;return $.includes(A.toLowerCase())?A:void 0},f$=($)=>{if($==null)return;try{return JSON.parse($)}catch(A){Y(A,"Failed to parse JSON",C);return}};var C$="context-request";class w$ extends Event{$;A;B;constructor($,A,B=!1){super(C$,{bubbles:!0,composed:!0});this.context=$;this.callback=A;this.subscribe=B}}var r=($)=>{let A=$.constructor,B=A.consumedContexts||[];setTimeout(()=>{for(let Z of B)$.dispatchEvent(new w$(Z,(W)=>$.set(String(Z),W)))});let K=A.providedContexts||[];if(!K.length)return!1;return $.addEventListener(C$,(Z)=>{let{context:W,callback:J}=Z;if(!K.includes(W)||!z(J))return;Z.stopPropagation(),J($.signals.get(String(W)))}),!0};class q extends HTMLElement{static registry=customElements;static states={};static observedAttributes;static consumedContexts;static providedContexts;static define($){try{if(q.registry.define($,this),Q)Y($,"Registered custom element")}catch(A){Y(A,`Failed to register custom element ${$}`,C)}}signals=new Map;self=new y(this);root=this.shadowRoot||this;debug=!1;attributeChangedCallback($,A,B){if(B===A)return;if(Q&&this.debug)Y(`${w(A)} => ${w(B)}`,`Attribute "${$}" of ${S(this)} changed`);this.set($,M(this,$,B,A))}connectedCallback(){if(Q){if(this.debug=this.hasAttribute("debug"),this.debug)Y(this,"Connected")}Object.entries(this.constructor.states).forEach(([$,A])=>{let B=z(A)?M(this,$,this.getAttribute($)??void 0,void 0):A;this.set($,B,!1)}),r(this)}disconnectedCallback(){if(Q&&this.debug)Y(this,"Disconnected")}adoptedCallback(){if(Q&&this.debug)Y(this,"Adopted")}has($){return this.signals.has($)}get($){let A=(K)=>!d(K)?K:z(K)?A(K()):I(K)?A(K.get()):K,B=A(this.signals.get($));if(Q&&this.debug)Y(B,`Get current value of state ${w($)} in ${S(this)}`);return B}set($,A,B=!0){let K;if(!this.signals.has($)){if(Q&&this.debug)K="Create";this.signals.set($,U(A,!0))}else if(B){let Z=this.signals.get($);if(I(A)){if(Q&&this.debug)K="Replace";if(this.signals.set($,A),L(Z))Z.set(V)}else if(L(Z)){if(Q&&this.debug)K="Update";Z.set(A)}else{Y(A,`Computed state ${w($)} in ${S(this)} cannot be set`,C);return}}else return;if(Q&&this.debug)Y(A,`${K} state ${w($)} in ${S(this)}`)}delete($){if(Q&&this.debug)Y($,`Delete state ${w($)} from ${S(this)}`);return this.signals.delete($)}first($){let A=this.root.querySelector($);return new y(this,A?[A]:[])}all($){return new y(this,Array.from(this.root.querySelectorAll($)))}}var R$=($,A=$)=>(B,K)=>D(()=>{K.dispatchEvent(new CustomEvent($,{detail:B.get(A),bubbles:!0}))}),N=($,A)=>(B,K)=>{let{read:Z,update:W}=A,J=Z(K);if(!z($)){let P=x($)&&x(J)?M(B,$,J):J;B.set($,P,!1)}D(()=>{let P=Z(K),j=z($)?$(P):B.get($);if(!Object.is(j,P)){if(j===null&&A.delete)A.delete(K);else if(j==null&&J)W(K,J);else if(j!=null)W(K,j)}})},O$=($,A)=>N(A,{read:()=>null,update:(B,K)=>A$(B,$,K)}),b$=($)=>N($,{read:(A)=>A!=null,update:(A,B)=>B?B$(A):Promise.resolve(null)}),k$=($)=>N($,{read:(A)=>A.textContent,update:(A,B)=>K$(A,B)}),_$=($,A=$)=>N(A,{read:(B)=>B[$],update:(B,K)=>B[$]=K}),E$=($,A=$)=>N(A,{read:(B)=>B.getAttribute($),update:(B,K)=>Z$(B,$,K),delete:(B)=>W$(B,$)}),h$=($,A=$)=>N(A,{read:(B)=>B.hasAttribute($),update:(B,K)=>J$(B,$,K)}),g$=($,A=$)=>N(A,{read:(B)=>B.classList.contains($),update:(B,K)=>Y$(B,$,K)}),d$=($,A=$)=>N(A,{read:(B)=>B.style.getPropertyValue($),update:(B,K)=>j$(B,$,K),delete:(B)=>Q$(B,$)});export{r as useContext,N as updateElement,g$ as toggleClass,h$ as toggleAttribute,U as toSignal,_ as state,k$ as setText,d$ as setStyle,_$ as setProperty,E$ as setAttribute,b$ as removeElement,M as parse,Y as log,L as isState,I as isSignal,X as enqueue,R$ as emit,D as effect,O$ as createElement,G as computed,t as batch,T$ as asString,M$ as asNumber,f$ as asJSON,y$ as asInteger,D$ as asEnum,q$ as asBoolean,e as animationFrame,V as UNSET,q as UIElement,y as UI,H$ as LOG_WARN,U$ as LOG_INFO,C as LOG_ERROR,z$ as LOG_DEBUG};
