+++ src/core/ui.d.ts
@@ -1,16 +1,15 @@
 import { type Signal } from '@zeix/cause-effect';
 import { UIElement } from '../ui-element';
 type StateLike<T> = PropertyKey | Signal<T> | ((v?: T) => T);
-type Factory<T> = (element: Element, index: number) => T;
-type FactoryOrValue<T> = T | Factory<T>;
-type StateLikeOrStateLikeFactory<T> = FactoryOrValue<StateLike<T>>;
-type EventListenerOrEventListenerFactory = FactoryOrValue<EventListenerOrEventListenerObject>;
+type ValueOrFactory<T> = T | ((element: Element, index: number) => T);
+type StateLikeOrStateLikeFactory<T> = ValueOrFactory<StateLike<T>>;
+type EventListenerOrEventListenerFactory = ValueOrFactory<EventListenerOrEventListenerObject>;
 declare class UI<T extends Element> {
     readonly host: UIElement;
     readonly targets: T[];
     constructor(host: UIElement, targets?: T[]);
-    on(event: string, listener: EventListenerOrEventListenerFactory): UI<T>;
-    off(event: string, listener: EventListenerOrEventListenerFactory): UI<T>;
+    on(type: keyof ElementEventMap, listener: EventListenerOrEventListenerFactory): UI<T>;
+    off(type: keyof ElementEventMap, listener: EventListenerOrEventListenerFactory): UI<T>;
     pass(states: Record<string, StateLikeOrStateLikeFactory<any>>): UI<T>;
     sync(...fns: ((host: UIElement, target: T, index: number) => void)[]): UI<T>;
 }
